% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_cat.R
\name{add_age_cat}
\alias{add_age_cat}
\alias{add_age_18_cat}
\title{Add categories of age}
\usage{
add_age_cat(
  df,
  age_col,
  breaks = seq(14, 100, by = 5),
  labels = NULL,
  int_undefined = c(-999, 999),
  char_undefined = "undefined",
  new_colname = NULL
)

add_age_18_cat(
  df,
  age_col,
  int_undefined = c(-999, 999),
  char_undefined = "undefined",
  new_colname = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{age_col}{The column name to recategorize.}

\item{breaks}{A vector of cut points.}

\item{labels}{A vector of labels. If NULL, the labels will be the breaks.}

\item{int_undefined}{A vector of values undefined (such as -999, 999) to replace by char_undefined.}

\item{char_undefined}{A character to replace int_undefined values, often values corresponding to Don't know of Pre.}

\item{new_colname}{The name of the new column. If NULL, it adds "_cat" to the age_col (or "_18_cat for \code{add_age_18_cat()}).}
}
\description{
\code{\link[=add_age_cat]{add_age_cat()}} adds the age categories and \code{\link[=add_age_18_cat]{add_age_18_cat()}} adds the age categories and a dummy variable for below and above 18.
}
\section{Details}{

\itemize{
\item \code{\link[=add_age_18_cat]{add_age_18_cat()}} adds two columns to the dataframe: one with the categories (minor, major) and one with a dummy variable for below and above 18.
}
}

