% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_received_assistance.R
\name{add_barriers_assistance_any}
\alias{add_barriers_assistance_any}
\title{Any barriers to accessing humanitarian assistance encountered}
\usage{
add_barriers_assistance_any(
  df,
  barriers_assistance = "aap_barriers_assistance",
  none = "none",
  pnta = "pnta",
  dnk = "dnk"
)
}
\arguments{
\item{df}{The input data frame}

\item{col_barrier}{The name of the column that contains the information about the barriers to accessing humanitarian assistance.}

\item{choice_none}{The choice value that indicates that no barriers were encountered.}

\item{choice_pnta}{The value value that indicates that the person prefers not to answer.}

\item{choice_dnk}{The value value that indicates that the person does not know.}
}
\value{
The data frame with the new variable \code{aap_barriers_assistance_any_barrier}
}
\description{
This function recodes any barriers to accessing humanitarian assistance encountered from the barriers to accessing humanitarian assistance by type indicator.
}
