% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_child_sep_cat.R
\name{add_child_sep_cat}
\alias{add_child_sep_cat}
\title{Add child separation categories}
\usage{
add_child_sep_cat(
  df,
  child_sep = "prot_child_sep",
  child_sep_yes = "yes",
  child_sep_no = "no",
  child_sep_undefined = c("pnta", "dnk"),
  child_sep_reason = "prot_child_sep_reason",
  child_sep_reason_non_severe = c("left_study"),
  child_sep_reason_severe = c("left_employment", "left_married"),
  child_sep_reason_very_severe = c("left_armed_groups", "kidnapped", "missing",
    "detained", "stayed_in_origin", "separated_displacement"),
  child_sep_reason_undefined = c("other", "dnk", "pnta"),
  sep = "/"
)
}
\arguments{
\item{df}{A data frame.}

\item{child_sep}{Child separation yes/no question.}

\item{child_sep_yes}{Value for yes.}

\item{child_sep_no}{Value for no.}

\item{child_sep_undefined}{Vector of undefined responses.}

\item{child_sep_reason}{Follow-up question for the reasons of child separation.}

\item{child_sep_reason_non_severe}{Values of non-severe reasons.}

\item{child_sep_reason_severe}{Values of severe reasons.}

\item{child_sep_reason_very_severe}{Values of very severe reasons.}

\item{child_sep_reason_undefined}{Values of undefined reasons.}

\item{sep}{Separator for the child_sep_reason columns.}
}
\description{
Add child separation categories
}
