% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy_from_dummies.R
\name{add_dummy_from_dummies}
\alias{add_dummy_from_dummies}
\title{Add a dummy variable from dummies}
\usage{
add_dummy_from_dummies(
  df,
  dummy_name = "dummy_from_dummies",
  ...,
  na_to_zero = FALSE
)
}
\arguments{
\item{df}{A data frame}

\item{dummy_name}{The new dummy column name}

\item{...}{Dummy columns from df, must be in set c(0,1).}

\item{na_to_zero}{Boolean. Mutate NA values to zero.}
}
\value{
An update data frame with a new dummy column.
}
\description{
Add a dummy variable from dummies
}
