% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_in_set.R
\name{add_dummy_from_set}
\alias{add_dummy_from_set}
\title{Add a dummy variable when any string from a set of strings}
\usage{
add_dummy_from_set(df, col, dummy_name, set, na_to_zero = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{col}{A column}

\item{dummy_name}{The new dummy column name}

\item{set}{A vector of strings to detect from, such as c("string1", "string2")}

\item{na_to_zero}{Boolean. Convert NAs to zeros or not.}
}
\value{
An update data frame with a new dummy column.
}
\description{
Add a dummy variable when any string from a set of strings
}
