% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edu_access_d.R
\name{add_edu_access_d}
\alias{add_edu_access_d}
\title{Add education access dummy}
\usage{
add_edu_access_d(
  df,
  edu_access = "edu_access",
  yes = "yes",
  no = "no",
  undefined = c("pnta", "dnk"),
  schooling_age_dummy = "ind_age_5_18"
)
}
\arguments{
\item{df}{A data frame of individual-level data.}

\item{edu_access}{Column name for education access}

\item{yes}{Value for yes}

\item{no}{Value for no}

\item{undefined}{Vector of undefined responses}

\item{schooling_age_dummy}{Column name for the dummy variable of schooling age}
}
\description{
Add education access dummy
}
