% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expenditure_type_prop_freq.R
\name{add_expenditure_type_prop_freq}
\alias{add_expenditure_type_prop_freq}
\title{Add Frequent Expenditure Type Amount as Proportions of Total Frequent Expenditure}
\usage{
add_expenditure_type_prop_freq(
  df,
  cm_expenditure_frequent_food = "cm_expenditure_frequent_food",
  cm_expenditure_frequent_rent = "cm_expenditure_frequent_rent",
  cm_expenditure_frequent_water = "cm_expenditure_frequent_water",
  cm_expenditure_frequent_nfi = "cm_expenditure_frequent_nfi",
  cm_expenditure_frequent_utilities = "cm_expenditure_frequent_utilities",
  cm_expenditure_frequent_fuel = "cm_expenditure_frequent_fuel",
  cm_expenditure_frequent_transportation = "cm_expenditure_frequent_transportation",
  cm_expenditure_frequent_communication = "cm_expenditure_frequent_communication",
  cm_expenditure_frequent_other = "cm_expenditure_frequent_other"
)
}
\arguments{
\item{df}{A data frame containing frequent expenditure data.}

\item{cm_expenditure_frequent_food}{Column name for food items expenditure amount.}

\item{cm_expenditure_frequent_rent}{Column name for rent or shelter and/or land expenditure amount.}

\item{cm_expenditure_frequent_water}{Column name for water expenditure amount from all sources combined.}

\item{cm_expenditure_frequent_nfi}{Column name for non-food household items expenditure amount for regular purchase, such as hygiene items, lightbulbs, etc.}

\item{cm_expenditure_frequent_utilities}{Column name for utilities expenditure amount, such as electricity or gas connections, etc.}

\item{cm_expenditure_frequent_fuel}{Column name for fuel expenditure amount for cooking, for vehicles, etc.}

\item{cm_expenditure_frequent_transportation}{Column name for transportation expenditure amount, not including vehicle fuel.}

\item{cm_expenditure_frequent_communication}{Column name for communications expenditure amount, such as phone airtime, Internet costs, etc.}

\item{cm_expenditure_frequent_other}{Column name for all other frequent expenditures.}
}
\value{
A data frame with additional columns:
\itemize{
\item cm_expenditure_frequent_total: The total frequent expenditure amount.
\item cm_expenditure_frequent_food_prop: Proportion of food expenditure.
\item cm_expenditure_frequent_rent_prop: Proportion of rent expenditure.
\item cm_expenditure_frequent_water_prop: Proportion of water expenditure.
\item cm_expenditure_frequent_nfi_prop: Proportion of non-food items expenditure.
\item cm_expenditure_frequent_utilities_prop: Proportion of utilities expenditure.
\item cm_expenditure_frequent_fuel_prop: Proportion of fuel expenditure.
\item cm_expenditure_frequent_transportation_prop: Proportion of transportation expenditure.
\item cm_expenditure_frequent_communication_prop: Proportion of communication expenditure.
\item cm_expenditure_frequent_other_prop: Proportion of other frequent expenditures.
}
}
\description{
This function calculates the proportion of each frequent expenditure type
relative to the total frequent expenditure. It adds new columns to the input
data frame representing these proportions.

Prerequisite function:
\itemize{
\item \code{\link[=add_expenditure_type_zero_freq]{add_expenditure_type_zero_freq()}}
}
}
