% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_income_source_cat.R
\name{add_income_source_cat}
\alias{add_income_source_cat}
\title{Add income source categories, count, and top 3}
\usage{
add_income_source_cat(
  df,
  emergency = c("cm_income_source_assistance_n", "cm_income_source_support_friends_n",
    "cm_income_source_donation_n"),
  unstable = c("cm_income_source_casual_n", "cm_income_source_social_benefits_n",
    "cm_income_source_rent_n", "cm_income_source_remittances_n"),
  stable = c("cm_income_source_salaried_n", "cm_income_source_own_business_n",
    "cm_income_source_own_production_n"),
  other = "cm_income_source_other_n",
  id_col = "uuid"
)
}
\arguments{
\item{df}{A data frame.}

\item{emergency}{The names of the columns that contain emergency income sources.}

\item{unstable}{The names of the columns that contain unstable income sources.}

\item{stable}{The names of the columns that contain stable income sources.}

\item{other}{The name of the column that contains other income sources.}

\item{id_col}{The name of the column that contains the unique identifier.}
}
\description{
Add income source categories, count, and top 3
}
