% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_interval.R
\name{add_quantile_interval}
\alias{add_quantile_interval}
\title{Add Quantile Intervals to Income Variable}
\usage{
add_quantile_interval(
  df,
  vars,
  cut_offs = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  weight = NULL
)
}
\arguments{
\item{df}{A dataframe containing the income variables.}

\item{vars}{The variable or variables with income integer values.}

\item{cut_offs}{The cut-offs for computation of quantiles. The default corresponds to the quantiles of the distribution, based on the weighted median.}

\item{weight}{The sampling weights column name. If NULL, unweighted quantiles are calculated.}
}
\value{
A dataframe with additional columns:
\itemize{
\item *_qtl: For each input variable, a new column is added with the quantile interval classification. The naming format is 'original_variable_name'_qtl.
}
}
\description{
This function calculates weighted income quantiles and classifies households according to these. It can handle multiple income variables and applies weighted or unweighted quantiles based on the input.
}
