% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_roster_wgq_ss.R
\name{add_roster_wgq_ss}
\alias{add_roster_wgq_ss}
\title{Add summarized individual WG-SS data to the household data}
\usage{
add_roster_wgq_ss(
  roster,
  hh,
  id_col = "uuid",
  wgq_disability = "wgq_disability",
  wgq_disability_3 = "wgq_disability_3",
  wgq_disability_4 = "wgq_disability_4",
  wgq_vision_3 = "wgq_vision_3",
  wgq_hearing_3 = "wgq_hearing_3",
  wgq_mobility_3 = "wgq_mobility_3",
  wgq_cognition_3 = "wgq_cognition_3",
  wgq_self_care_3 = "wgq_self_care_3",
  wgq_communication_3 = "wgq_communication_3",
  wgq_vision_4 = "wgq_vision_4",
  wgq_hearing_4 = "wgq_hearing_4",
  wgq_mobility_4 = "wgq_mobility_4",
  wgq_cognition_4 = "wgq_cognition_4",
  wgq_self_care_4 = "wgq_mobility_4",
  wgq_communication_4 = "wgq_communication_4"
)
}
\arguments{
\item{roster}{A data frame of individual-level data.}

\item{hh}{A data frame of household-level data.}

\item{id_col}{Survey unique identifier column (must be present in both roster and hh).}

\item{wgq_disability}{Dummy from roster. One group from the SS was reported as A lot of difficulty or Cannot do at all.}

\item{wgq_disability_3}{Dummy from roster. One group from the SS was reported as A lot of difficulty.}

\item{wgq_disability_4}{Dummy from roster. One group from the SS was reported as Cannot do at all.}

\item{wgq_vision_3}{Dummy from roster. Group vision was reported as A lot of difficulty.}

\item{wgq_hearing_3}{Dummy from roster. Group hearing was reported as A lot of difficulty.}

\item{wgq_mobility_3}{Dummy from roster. Group mobility was reported as A lot of difficulty.}

\item{wgq_cognition_3}{Dummy from roster. Group cognition was reported as A lot of difficulty.}

\item{wgq_self_care_3}{Dummy from roster. Group self care was reported as A lot of difficulty.}

\item{wgq_communication_3}{Dummy from roster. Group communication was reported as A lot of difficulty.}

\item{wgq_vision_4}{Dummy from roster. Group vision was reported as Cannot do at all.}

\item{wgq_hearing_4}{Dummy from roster. Group hearing was reported as Cannot do at all.}

\item{wgq_mobility_4}{Dummy from roster. Group mobility was reported as Cannot do at all.}

\item{wgq_cognition_4}{Dummy from roster. Group cognition was reported as Cannot do at all.}

\item{wgq_self_care_4}{Dummy from roster. Group self care was reported as Cannot do at all.}

\item{wgq_communication_4}{Dummy from roster. Group communication was reported as Cannot do at all.}
}
\value{
hh with fifteen new columns: each component dummy for levels 3 and 4, disability level 3, level 4, and level 3 and 4 together (summarized by household).
}
\description{
Add summarized individual WG-SS data to the household data
}
\details{
To be used after \code{roster_wgq_ss()}
}
