% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shelter_issue_cat.R
\name{add_shelter_issue_cat}
\alias{add_shelter_issue_cat}
\title{Add Number of Shelter Issues and Related Category}
\usage{
add_shelter_issue_cat(
  df,
  shelter_issue = "snfi_shelter_issue",
  none = "none",
  issues = c("lack_privacy", "lack_space", "temperature", "ventilation", "leak", "lock",
    "lack_lighting", "difficulty_move"),
  undefined = c("dnk", "pnta"),
  other = c("other"),
  sep = "/"
)
}
\arguments{
\item{df}{A data frame containing shelter issue data.}

\item{shelter_issue}{Component column: Shelter issues.}

\item{none}{Response code for no issue.}

\item{issues}{Character vector of issues.}

\item{undefined}{Character vector of undefined responses codes (e.g. "Prefer not to answer").}

\item{other}{Character vector of other responses codes (e.g. "Other").}

\item{sep}{Separator for the binary columns.}
}
\value{
A data frame with additional columns:
\item{snfi_shelter_issue_n}{Count of shelter issues}
\item{snfi_shelter_issue_cat}{Categorized shelter issues: "none", "undefined", "other", "1_to_3", "4_to_6", or "7_to_8"}
}
\description{
This function calculates the number of shelter issues and categorizes them based on predefined thresholds. It also handles undefined and other responses.
}
