% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shelter_type_cat.R
\name{add_shelter_type_cat}
\alias{add_shelter_type_cat}
\title{Combine and Recode Shelter Types}
\usage{
add_shelter_type_cat(
  df,
  shelter_type = "snfi_shelter_type",
  sl_none = "none",
  sl_collective_center = "collective_center",
  sl_undefined = "pnta",
  shelter_type_individual = "snfi_shelter_type_individual",
  adequate = c("house", "apartment", "tent"),
  inadequate = c("makeshift", "unfinished_building"),
  undefined = c("pnta", "other", "dnk")
)
}
\arguments{
\item{df}{A data frame containing shelter type information.}

\item{shelter_type}{Component column: Shelter type categories.}

\item{sl_none}{Character vector of responses codes for none/sleeping in the open that are skipped.}

\item{sl_collective_center}{Character vector of responses codes for collective center that are skipped.}

\item{sl_undefined}{Character vector of undefined responses codes (e.g. "Prefer not to answer") that are skipped.}

\item{shelter_type_individual}{Component column: Individual shelter types.}

\item{adequate}{Character vector of responses codes for adequate shelter types.}

\item{inadequate}{Character vector of responses codes for inadequate shelter types.}

\item{undefined}{Character vector of responses codes for undefined shelter types.}
}
\value{
A data frame with an additional column:
\item{snfi_shelter_type_cat}{Categorized shelter type: "none", "inadequate", "adequate", or "undefined"}
}
\description{
This function combines both shelter types questions and recodes the type of shelter into categories such as "none", "inadequate", "adequate", or "undefined".
}
