% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_top3_expenditure_type_infreq.R
\name{add_top3_expenditure_type_infreq}
\alias{add_top3_expenditure_type_infreq}
\title{Rank top 3 infrequent expenditure types}
\usage{
add_top3_expenditure_type_infreq(
  df,
  expenditure_infreq_types = c("cm_expenditure_infrequent_shelter",
    "cm_expenditure_infrequent_nfi", "cm_expenditure_infrequent_health",
    "cm_expenditure_infrequent_education", "cm_expenditure_infrequent_debt",
    "cm_expenditure_infrequent_other"),
  id_col = "uuid"
)
}
\arguments{
\item{df}{A data frame.}

\item{expenditure_infreq_types}{A character vector. The names of the columns that contain the amount of infrequent expenditures types.}

\item{id_col}{The name of the column that contains the unique identifier.}
}
\description{
Rank top 3 infrequent expenditure types
}
