% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{are_cols_integer}
\alias{are_cols_integer}
\title{Stop statement values are not integer}
\usage{
are_cols_integer(df, cols)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}
}
\value{
A stop statement
}
\description{
This stop statement checks that specified columns are integer or numeric with all integer values (to not be too restrictive as the aim is not much to ensure integer values but not integer type).
}
