% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{are_cols_numeric}
\alias{are_cols_numeric}
\title{Stop statement values are not numeric}
\usage{
are_cols_numeric(df, cols)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}
}
\value{
A stop statement
}
\description{
Stop statement values are not numeric
}
