% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{are_values_in_range}
\alias{are_values_in_range}
\title{Stop statement values are not in range}
\usage{
are_values_in_range(df, cols, lower = 0, upper = 7)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\value{
A stop statement
}
\description{
Stop statement values are not in range
}
