% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_health.R
\name{comp_health}
\alias{comp_health}
\title{Health composite - add score and dummy for in need}
\usage{
comp_health(
  df,
  ind_healthcare_needed_no_n = "health_ind_healthcare_needed_no_n",
  ind_healthcare_needed_yes_unmet_n = "health_ind_healthcare_needed_yes_unmet_n",
  ind_healthcare_needed_yes_met_n = "health_ind_healthcare_needed_yes_met_n",
  wgq_dis = FALSE,
  ind_healthcare_needed_no_wgq_dis_n = "health_ind_healthcare_needed_no_wgq_dis_n",
  ind_healthcare_needed_yes_unmet_wgq_dis_n =
    "health_ind_healthcare_needed_yes_unmet_wgq_dis_n",
  ind_healthcare_needed_yes_met_wgq_dis_n =
    "health_ind_healthcare_needed_yes_met_wgq_dis_n"
)
}
\arguments{
\item{df}{A data frame.}

\item{ind_healthcare_needed_no_n}{Column name for the number of individuals who did not need to access healthcare.}

\item{ind_healthcare_needed_yes_unmet_n}{Column name for the number of individuals who needed to access healthcare but did not.}

\item{ind_healthcare_needed_yes_met_n}{Column name for the number of individuals who needed to access healthcare and did.}

\item{wgq_dis}{Logical. If TRUE, the function will consider the disability columns.}

\item{ind_healthcare_needed_no_wgq_dis_n}{Column name for the number of individuals who did not need to access healthcare and have a disability.}

\item{ind_healthcare_needed_yes_unmet_wgq_dis_n}{Column name for the number of individuals who needed to access healthcare but did not and have a disability.}

\item{ind_healthcare_needed_yes_met_wgq_dis_n}{Column name for the number of individuals who needed to access healthcare and did and have a disability.}
}
\description{
Health composite - add score and dummy for in need
}
