% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinking_water_quantity.R
\name{drinking_water_quantity}
\alias{drinking_water_quantity}
\title{Drinking water quantity classification - 5-point scale}
\usage{
drinking_water_quantity(
  df,
  drinking_water_quantity = "water_source",
  level_codes = c("always", "often", "sometimes", "rarely", "never"),
  class_colname = "drinking_water_quantity_class"
)
}
\arguments{
\item{df}{A data frame.}

\item{drinking_water_quantity}{Component column: Water source types.}

\item{level_codes}{Character vector of responses codes, including first in the following order: "Always (more than 20 times)", "Often (11-20 times)", "Sometimes (3–10 times)", "Rarely (1–2 times)", "Never (0 times)", e.g. c("always", "often", "sometimes", "rarely", "never").}

\item{class_colname}{The new column name for the classification column. Default to "drinking_water_quantity_class".}
}
\value{
One new column: a 5-point scale from 1 to 5 (drinking_water_quantity_class).
}
\description{
\code{drinking_water_quantity()} recodes the quantity of drinking availibility on a 5-point scale.
}
