% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_raw_data}
\alias{dummy_raw_data}
\title{Some MSNA dummy data
A list of datasets with "main" for household-level, "roster" for demographics, "edu_ind" for individual education variables, "health_ind" for individual health variables, and "nut_ind" for individual nutrition variables.}
\format{
A list of tibbles
}
\usage{
dummy_raw_data
}
\description{
Some MSNA dummy data
A list of datasets with "main" for household-level, "roster" for demographics, "edu_ind" for individual education variables, "health_ind" for individual health variables, and "nut_ind" for individual nutrition variables.
}
\keyword{datasets}
