% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm_cell.R
\name{fcm_cell}
\alias{fcm_cell}
\title{FEWS NET Food Consumption Matrix cell}
\usage{
fcm_cell(
  df,
  fcs_cat = "fcs_cat",
  fcs_levels = c("Acceptable", "Borderline", "Poor"),
  hhs_cat = "hhs_cat",
  hhs_levels = c("None", "Little", "Moderate", "Severe", "Very severe"),
  rcsi_cat = "rcsi_cat",
  rcsi_levels = c("No to Low", "Medium", "Severe")
)
}
\arguments{
\item{df}{A data frame.}

\item{fcs_cat}{Food Consumption Score column.}

\item{fcs_levels}{Food Consumption Score categories - in that order: "Poor", "Borderline", and "Acceptable".}

\item{hhs_cat}{Household Hunger Scale column.}

\item{hhs_levels}{Household Hunger Scale categories - in that order: "None", "Little", "Moderate", "Severe", and "Very severe".}

\item{rcsi_cat}{reduced Coping Strategies Index column.}

\item{rcsi_levels}{reduced Coping Strategies Index categories - in that order: "No to Low", "Medium", and "Severe".}
}
\value{
The FEWS NET Food Consumption Cell number.
}
\description{
FEWS NET Food Consumption Matrix cell
}
