% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcp.R
\name{fcp}
\alias{fcp}
\title{FC phase - Food Consumption phase according to the FCM}
\usage{
fcp(df, fcm_cell = "fcm_cell")
}
\arguments{
\item{df}{A data frame.}

\item{fcm_cell}{Food consumption matrix cell - 1 to 45.}
}
\value{
Two columns: the food consumption phase number (fcp) and the category (fcp_cat)
}
\description{
FC phase - Food Consumption phase according to the FCM
}
