% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fies.R
\name{fies_raw_score}
\alias{fies_raw_score}
\title{FIES - Food Insecurity Experience Scale analysis}
\usage{
fies_raw_score(
  df,
  fies_worried = "fies_worried",
  fies_healthy = "fies_healthy",
  fies_fewfoods = "fies_fewfoods",
  fies_skipped = "fies_skipped",
  fies_ateless = "fies_ateless",
  fies_ranout = "fies_ranout",
  fies_hungry = "fies_hungry",
  fies_wholeday = "fies_wholeday",
  level_codes = c("yes", "no", "do_not_know", "prefer_not_to_answer")
)
}
\arguments{
\item{df}{A data frame}

\item{fies_worried}{Component column: Worried about not having enough food to eat.}

\item{fies_healthy}{Component column: Unable to eat healthy or nutritious food.}

\item{fies_fewfoods}{Component column: Ate only a few kinds of foods.}

\item{fies_skipped}{Component column: Skip a meal.}

\item{fies_ateless}{Component column: Ate less than thought to should.}

\item{fies_ranout}{Component column: Ran out of food.}

\item{fies_hungry}{Component column: Hungry without eating.}

\item{fies_wholeday}{Component column: No eating for a whole day.}

\item{level_codes}{Character vector of responses codes, including first in the following order: "Yes", "No", "Do not know", "Prefer not to answer", e.g., c("yes", "no", "do_not_know", "prefer_not_to_answer")}
}
\value{
Nine new columns: each component dummy variable if "Yes" (fies_*_d) and FIES raw score (fiew_raw_score).
}
\description{
\code{fies_raw_score()} calculates the raw score of the FIES. For parameter estimates and assessment statistics with a Rasch model thanks to package \code{RM.weights}, see \code{impactR.analysis::fies_estimates()} (under construction).
}
\details{
THe recall period can be different (30 days or 12 months) depending on the research's objectives (food insecurity or population monitoring for instance).
}
