% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodsec.R
\name{foodsec}
\alias{foodsec}
\title{Calculate main food security indicators}
\usage{
foodsec(
  df,
  fcs_cereal = "fcs_cereal",
  fcs_legumes = "fcs_legumes",
  fcs_dairy = "fcs_dairy",
  fcs_meat = "fcs_meat",
  fcs_veg = "fcs_veg",
  fcs_fruit = "fcs_fruit",
  fcs_oil = "fcs_oil",
  fcs_sugar = "fcs_sugar",
  fcs_cat = "normal",
  hhs_lev1_nofoodhh = "hhs_lev1_nofoodhh",
  hhs_lev2_nofoodhh = "hhs_lev2_nofoodhh",
  hhs_lev1_sleephungry = "hhs_lev1_sleephungry",
  hhs_lev2_sleephungry = "hhs_lev2_sleephungry",
  hhs_lev1_alldaynight = "hhs_lev1_alldaynight",
  hhs_lev2_alldaynight = "hhs_lev2_alldaynight",
  hhs_level1_codes = c("yes", "no"),
  hhs_level2_codes = c("rarely", "sometimes", "often"),
  rcsi_lesspreferred = "rcsi_lesspreferred",
  rcsi_borrowfood = "rcsi_borrowfood",
  rcsi_limitportion = "rcsi_limitportion",
  rcsi_restrict = "rcsi_restrict",
  rcsi_reducemeals = "rcsi_reducemeals",
  lcsi_stress_1 = "lcsi_stress_1",
  lcsi_stress_2 = "lcsi_stress_2",
  lcsi_stress_3 = "lcsi_stress_3",
  lcsi_stress_4 = "lcsi_stress_4",
  lcsi_crisis_1 = "lcsi_crisis_1",
  lcsi_crisis_2 = "lcsi_crisis_2",
  lcsi_crisis_3 = "lcsi_crisis_3",
  lcsi_emergency_1 = "lcsi_emergency_1",
  lcsi_emergency_2 = "lcsi_emergency_2",
  lcsi_emergency_3 = "lcsi_emergency_3",
  lcsi_level_codes = c("yes", "exhausted", "no_need", "not_applicable")
)
}
\arguments{
\item{df}{Data}

\item{fcs_cereal}{Cereals component column}

\item{fcs_legumes}{Legumes component column}

\item{fcs_dairy}{Dairy component column}

\item{fcs_meat}{Meat component column}

\item{fcs_veg}{Vegetables component column}

\item{fcs_fruit}{Fruit component column}

\item{fcs_oil}{Oil component column}

\item{fcs_sugar}{Sugar component column}

\item{fcs_cat}{Cut-offs categories, either "normal" (21.5/35) or "alternate" (28/42)}

\item{hhs_lev1_nofoodhh}{Component column: No food of any kind in the house}

\item{hhs_lev2_nofoodhh}{Follow-up frequency column}

\item{hhs_lev1_sleephungry}{Component column: Go to sleep hungry because there was not enough food}

\item{hhs_lev2_sleephungry}{Follow-up frequency column}

\item{hhs_lev1_alldaynight}{Component column: Go a whole day and night without eating}

\item{hhs_lev2_alldaynight}{Follow-up frequency column}

\item{hhs_level1_codes}{Character vector of at least "Yes" and "No" codes (in this order), e.g. c("yes", "no")}

\item{hhs_level2_codes}{Character vector of at least frequencies codes, in the following order: "Rarely", "Sometimes", "Often", e.g. c("rarely", "sometimes", "often")}

\item{rcsi_lesspreferred}{Component column: Rely on less preferred and less expensive food}

\item{rcsi_borrowfood}{Component column: Borrow food or rely on help from friends or relatives}

\item{rcsi_limitportion}{Component column: Limit portion size at mealtime}

\item{rcsi_restrict}{Component column: Restrict consumption by adults in order for small children to eat}

\item{rcsi_reducemeals}{Component column: Reduce the number of meals eaten in a day}

\item{lcsi_stress_1}{Component column: Stress strategy 1}

\item{lcsi_stress_2}{Component column: Stress strategy 2}

\item{lcsi_stress_3}{Component column: Stress strategy 3}

\item{lcsi_stress_4}{Component column: Stress strategy 4}

\item{lcsi_crisis_1}{Component column: Crisis strategy 1}

\item{lcsi_crisis_2}{Component column: Crisis strategy 2}

\item{lcsi_crisis_3}{Component column: Crisis strategy 3}

\item{lcsi_emergency_1}{Component column: Emergency strategy 1}

\item{lcsi_emergency_2}{Component column: Emergency strategy 2}

\item{lcsi_emergency_3}{Component column: Emergency strategy 3}

\item{lcsi_level_codes}{Character vector of at responses codes, in the following order: ""Yes", "No, exhausted", "No, no need", "No, not applicable", e.g. c("yes", "exhausted", "no_need", "not_applicable")}
}
\value{
All columns produced by the following \code{impactR} functions: \code{fcs()}, \code{hhs()}, \code{rcsi()}, \code{lcs()}, \code{fcm()}, \code{fcp()} and \code{fclp()}.
}
\description{
Calculate main food security indicators
}
