% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdds.R
\name{hdds}
\alias{hdds}
\title{HHDS - Household Dietary Diversity Score (standalone version)}
\usage{
hdds(
  df,
  hdds_cereal = "hdds_ceral",
  hdds_roots = "hdds_roots",
  hdds_legumes = "hdds_legumes",
  hdds_dairy = "hdds_dairy",
  hdds_meat = "hdds_meat",
  hdds_fish = "hdds_fish",
  hdds_egg = "hdds_egg",
  hdds_veg = "hdds_veg",
  hdds_fruits = "hdds_fruits",
  hdds_oil = "hdds_oil",
  hdds_sugar = "hdds_sugar",
  hdds_spices = "hdds_spices",
  level_codes = c("yes", "no", "do_not_know", "prefer_not_to_answer")
)
}
\arguments{
\item{df}{A data frame}

\item{hdds_cereal}{Cereals and grain component column}

\item{hdds_roots}{Roots and tubers component column}

\item{hdds_legumes}{Legumes, pulses, and nuts component column}

\item{hdds_dairy}{Dairy, milk, and others component column}

\item{hdds_meat}{Meat component column}

\item{hdds_fish}{Fish and shellfish component column}

\item{hdds_egg}{Eggs component column}

\item{hdds_veg}{Vegetables component column}

\item{hdds_fruits}{Fruits component column}

\item{hdds_oil}{Oil component column}

\item{hdds_sugar}{Sugar component column}

\item{hdds_spices}{Spices component column}

\item{level_codes}{Character vector of responses codes, including first in the following order: "Yes", "No", "Do not know", "Prefer not to answer", e.g., c("yes", "no", "do_not_know", "prefer_not_to_answer").}
}
\value{
Thirteen new columns: each component dummy variable if "Yes" (hdds_*_d) and HDDS score (hdds_score).
}
\description{
HHDS - Household Dietary Diversity Score (standalone version)
}
