% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwise.R
\name{hwise}
\alias{hwise}
\title{HWISE scores - Household Water Insecurity Experiences (HWISE) Scale}
\usage{
hwise(
  df,
  hwise_worry = "hwise_worry",
  hwise_interrupt = "hwise_interrupt",
  hwise_clothes = "hwise_clothes",
  hwise_plans = "hwise_plans",
  hwise_food = "hwise_food",
  hwise_hands = "hwise_hands",
  hwise_body = "hwise_body",
  hwise_drink = "hwise_drink",
  hwise_angry = "hwise_angry",
  hwise_sleep = "hwise_sleep",
  hwise_none = "hwise_none",
  hwise_shame = "hwise_shame",
  level_codes = c("never", "rarely", "sometimes", "often", "always", "do_not_know",
    "prefer_not_to_answer", "not_applicable")
)
}
\arguments{
\item{df}{A data frame}

\item{hwise_worry}{Component column: Worry.}

\item{hwise_interrupt}{Component column: Interrupt.}

\item{hwise_clothes}{Component column: Clothes.}

\item{hwise_plans}{Component column: Plans.}

\item{hwise_food}{Component column: Food.}

\item{hwise_hands}{Component column: Hands.}

\item{hwise_body}{Component column: Body.}

\item{hwise_drink}{Component column: Drink.}

\item{hwise_angry}{Component column: Angry.}

\item{hwise_sleep}{Component column: Sleep.}

\item{hwise_none}{Component column: None.}

\item{hwise_shame}{Component column: Shame.}

\item{level_codes}{Character vector of responses codes, including first in the following order: "Never (0 times)", "Rarely (1–2 times)", "Sometimes (3–10 times)", "Often (11-20 times)", "Always (more than 20 times)", "Do not know", "Prefer not to answer", "Not applicable", e.g. c("never", "rarely", "sometimes", "often", "always", "do_not_know", "prefer_not_to_answer", "not_applicable").}
}
\value{
Sixteen new columns: each component score (hwise_score_*), the overall score (hwise_score), dummy for water insecurity (hwise_insecure), the broader categories (hwise_cat), the detailed categories (hwise_cat_details).
}
\description{
HWISE scores - Household Water Insecurity Experiences (HWISE) Scale
}
