% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_value}
\alias{impute_value}
\alias{impute_median}
\title{Impute missing values}
\usage{
impute_value(df, vars, value)

impute_median(df, vars, group = NULL, weighted = FALSE, weight = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{A character vector of the variables to replace na with.}

\item{value}{The value to replace na with.}

\item{group}{A character vector of the grouping variables.}

\item{weighted}{A boolean indicating whether to use the weighted median or not.}

\item{weight}{The weight variable.}
}
\description{
\code{\link[=impute_value]{impute_value()}} replaces all missing values by a specific value and \code{\link[=impute_median]{impute_median()}} replaces all missing values by the (weighted) median of the variable.
}
