% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsi.R
\name{lcsi}
\alias{lcsi}
\title{LCSI - Livelihood Coping Strategy Index}
\usage{
lcsi(
  df,
  lcsi_stress_1 = "lcsi_stress_1",
  lcsi_stress_2 = "lcsi_stress_2",
  lcsi_stress_3 = "lcsi_stress_3",
  lcsi_stress_4 = "lcsi_stress_4",
  lcsi_crisis_1 = "lcsi_crisis_1",
  lcsi_crisis_2 = "lcsi_crisis_2",
  lcsi_crisis_3 = "lcsi_crisis_3",
  lcsi_emergency_1 = "lcsi_emergency_1",
  lcsi_emergency_2 = "lcsi_emergency_2",
  lcsi_emergency_3 = "lcsi_emergency_3",
  level_codes = c("yes", "exhausted", "no_need", "not_applicable")
)
}
\arguments{
\item{df}{A data frame}

\item{lcsi_stress_1}{Component column: Stress strategy 1}

\item{lcsi_stress_2}{Component column: Stress strategy 2}

\item{lcsi_stress_3}{Component column: Stress strategy 3}

\item{lcsi_stress_4}{Component column: Stress strategy 4}

\item{lcsi_crisis_1}{Component column: Crisis strategy 1}

\item{lcsi_crisis_2}{Component column: Crisis strategy 2}

\item{lcsi_crisis_3}{Component column: Crisis strategy 3}

\item{lcsi_emergency_1}{Component column: Emergency strategy 1}

\item{lcsi_emergency_2}{Component column: Emergency strategy 2}

\item{lcsi_emergency_3}{Component column: Emergency strategy 3}

\item{level_codes}{Character vector of responses codes, including first in the following order: Yes", "No, exhausted", "No, no need", "No, not applicable", e.g. c("yes", "exhausted", "no_need", "not_applicable")}
}
\value{
Fourteen new columns: each strategy recoded (lcsi_stress_\emph{, lcsi_crisis_}, lcsi_emergency_*), a dummy for each category (lcsi_stress, lcsi_crisis, lcsi_emergency), and the category (lcsi_cat).
}
\description{
LCSI - Livelihood Coping Strategy Index
}
