% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_cat.R
\name{num_cat}
\alias{num_cat}
\title{Add categories for a numeric variable}
\usage{
num_cat(
  df,
  num_col,
  breaks,
  labels = NULL,
  int_undefined = c(-999, 999),
  char_undefined = "Unkown",
  new_colname = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{num_col}{The column name to recategorize.}

\item{breaks}{A vector of cut points.}

\item{labels}{A vector of labels.}

\item{int_undefined}{A vector of values to replace by char_dontknow.}

\item{char_undefined}{A character to replace int_dontknow values.}

\item{new_colname}{The name of the new column.}
}
\value{
A dataframe with a new column.
}
\description{
Add categories for a numeric variable
}
