% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_top3_vars.R
\name{rank_top3_vars}
\alias{rank_top3_vars}
\title{Function to add top 3 columns out of numeric variables}
\usage{
rank_top3_vars(
  df,
  vars,
  new_colname_top1,
  new_colname_top2,
  new_colname_top3,
  id_col = "uuid"
)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{A character vector of numeric variable names.}

\item{new_colname_top1}{The new column name for the top variable.}

\item{new_colname_top2}{The new column name for the 2nd top variable.}

\item{new_colname_top3}{The new column name for the 3rd top variable.}

\item{id_col}{The column name for the unique identifier.}
}
\description{
In case of equality, this function takes precedence in the order of 'vars' as factors.
}
