% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-impactR4PHU.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{add_hhs}
\alias{add_fcs}
\alias{add_rcsi}
\alias{add_lcsi}
\alias{add_fcm_phase}
\alias{add_fclcm_phase}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{impactR4PHU}{\code{\link[impactR4PHU:add_fclcm_phase]{add_fclcm_phase()}}, \code{\link[impactR4PHU:add_fcm_phase]{add_fcm_phase()}}, \code{\link[impactR4PHU:add_fcs]{add_fcs()}}, \code{\link[impactR4PHU:add_hhs]{add_hhs()}}, \code{\link[impactR4PHU:add_lcsi]{add_lcsi()}}, \code{\link[impactR4PHU:add_rcsi]{add_rcsi()}}}
}}

