% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster_wgq_ss.R
\name{roster_wgq_ss}
\alias{roster_wgq_ss}
\title{Prepare dummy variables for each WG-SS component (individual data)}
\usage{
roster_wgq_ss(
  roster,
  ind_age = "ind_age",
  vision = " wgq_vision",
  hearing = "wgq_hearing",
  mobility = "wgq_mobility",
  cognition = "wgq_cognition",
  self_care = "wgq_self_care",
  communication = "wgq_communication",
  level_codes = c("no_difficulty", "some_difficulty", "lot_of_difficulty", "cannot_do")
)
}
\arguments{
\item{roster}{A data frame of individual-level data.}

\item{ind_age}{The individual age column.}

\item{vision}{Vision component column.}

\item{hearing}{Hearing component column.}

\item{mobility}{Mobility component column.}

\item{cognition}{Cognition component column.}

\item{self_care}{Self-care component column.}

\item{communication}{Communication component column.}

\item{level_codes}{Character vector of responses codes, including first in the following order: "No difficulty", "Some difficulty", "A lot of difficulty", "Cannot do at all", e.g. c("no_difficulty", "some_difficulty", "lot_of_difficulty", "cannot_do").}
}
\value{
Fifteen new columns: each component dummy for levels 3 and 4, disability level 3, level 4, and level 3 and 4 together.
}
\description{
Prepare dummy variables for each WG-SS component (individual data)
}
