% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_vars.R
\name{sum_vars}
\alias{sum_vars}
\title{Function to sum up columns row-wise.}
\usage{
sum_vars(
  df,
  vars,
  new_colname,
  imputation = "none",
  na_rm = FALSE,
  weight = "weight",
  value = 0,
  group = NULL
)
}
\arguments{
\item{df}{A dataframe}

\item{vars}{A character vector of the columns to sum.}

\item{new_colname}{A character vector of the new column name.}

\item{imputation}{The imputation method, either none (default), value, median or weighted median.}

\item{na_rm}{A boolean indicating whether to remove missing values when summing across columns.}

\item{weight}{The weight variable to calculate weighted means or medians.}

\item{value}{The value to replace missing values with if imputation is "value".}

\item{group}{A character vector of the grouping variables.}
}
\description{
Function to sum up columns row-wise.
}
