% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_to_sl.R
\name{value_to_sl}
\alias{value_to_sl}
\title{Add a specific value to variables that were skipped}
\usage{
value_to_sl(df, var, undefined = NULL, sl_vars, sl_value, suffix = "")
}
\arguments{
\item{df}{A data frame.}

\item{var}{The name of the variable that is used for the skip logic.}

\item{undefined}{A character vector of values that are considered as "undefined" and should not be replaced in follow-up variables.}

\item{sl_vars}{A character vector of variable names which were skipped and should be replaced with a value.}

\item{sl_value}{The value that should be added to the skipped variables.}

\item{suffix}{A suffix to add to the new variable names. Default to no suffix (empty string).}
}
\description{
Add a specific value to variables that were skipped
}
