% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_source.R
\name{water_source}
\alias{water_source}
\title{Water source classification - 5-point scale}
\usage{
water_source(
  df,
  water_source = "water_source",
  water_source_improved_codes = c("protected_well", "public_tap"),
  water_source_unimproved_codes = c("unprotected_well", "unprotected_spring"),
  water_source_surface_water_codes = c("lake", "river"),
  water_source_na_codes = c("other"),
  time_to_fetch = "time_to_fetch_water",
  time_to_fetch_above_threshold_codes = c("30mins_to_59mins", "1hour_above"),
  time_to_fetch_below_threshold_codes = c("5mins_to_14mins", "15mins_to_29mins"),
  time_to_fetch_premises_codes = c("premises"),
  time_to_fetch_na_codes = c("other")
)
}
\arguments{
\item{df}{A data frame.}

\item{water_source}{Component column: Water source types.}

\item{water_source_improved_codes}{Character vector of responses codes, such as "Protected well" or "Public tap", e.g., c("protected_well", "public_tap").}

\item{water_source_unimproved_codes}{Character vector of responses codes, such as "Unprotected well" or "Unprotected spring", e.g., c("unprotected_well", "unprotected_spring").}

\item{water_source_surface_water_codes}{Character vector of responses codes, such as "Lake" or "River, e.g., c("lake", "river").}

\item{water_source_na_codes}{Character vector of responses codes, that do not fit any category, e.g., c("other").}

\item{time_to_fetch}{Component column: Time to fetch water.}

\item{time_to_fetch_above_threshold_codes}{Character vector of responses codes, such as "30 minutes to 59 minutes" or "1 hour and above", e.g., c("30mins_to_59mins", "1hour_above").}

\item{time_to_fetch_below_threshold_codes}{Character vector of responses codes, such as "5 to 14 minutes" or "15 minutes to 29 minutes", e.g., c("5mins_to_14mins", "15mins_to_29mins").}

\item{time_to_fetch_premises_codes}{Character vector of responses codes, such as "On premises", e.g., c("premises").}

\item{time_to_fetch_na_codes}{Character vector of responses codes, that do not fit any category, e.g., c("other").}
}
\value{
Three new columns: a recoded column of water sources between improved, unimproved and surface water (water_source_recoded), a recoded column of times to fetch water according to the chosen thresholds (time_to_fetch_recoded), a 5-point scale from 1 to 5 (water_source_class).
}
\description{
\code{water_source()} recodes the types of water sources, the time to fetch water according to a chosen threshold, and classify each household/individual on a 5-point scale.
}
\section{Details on the 5-point scale}{


The classification is computed as follows:
\itemize{
\item Level 5: Surface water;
\item Level 4: Unimproved water source;
\item Level 3: Improved water source with more than 30 minutes return time;
\item Level 2: Improved water source within 30 minutes return time;
\item Level 1: Improved water source on premises.
}
}

