% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_edu.R
\name{add_comp_edu}
\alias{add_comp_edu}
\title{Calculate Education Sectoral Composite Score and Need Indicators}
\usage{
add_comp_edu(
  df,
  schooling_age_n = "edu_schooling_age_n",
  no_access_n = "edu_no_access_n",
  barrier_protection_n = "edu_barrier_protection_n",
  attack_n = "edu_disrupted_attack_n",
  hazards_n = "edu_disrupted_hazards_n",
  displaced_n = "edu_disrupted_displaced_n",
  teacher_n = "edu_disrupted_teacher_n"
)
}
\arguments{
\item{df}{A data frame.}

\item{schooling_age_n}{Column name for the number of children of schooling age.}

\item{no_access_n}{Column name for the number of children with no access to education.}

\item{barrier_protection_n}{Column name for the number of children with barriers to protection.}

\item{attack_n}{Column name for the number of children with disrupted education due to the school being occupied by armed groups}

\item{hazards_n}{Column name for the number of children with disrupted education due to hazards.}

\item{displaced_n}{Column name for the number of children with disrupted education due to a recent displacement.}

\item{teacher_n}{Column name for the number of children with disrupted education due to teachers' absence.}
}
\value{
A data frame with additional columns:
\itemize{
\item comp_edu_score_disrupted: Score for disrupted education (1-4).
\item comp_edu_score_attendance: Score for attendance and barriers (1-4)
\item comp_edu_score: Total education composite score (max of disrupted and attendance scores).
\item comp_edu_in_need: Binary indicator for being in need of educational assistance.
\item comp_edu_in_severe_need: Binary indicator for being in severe need of educational assistance.
}
}
\description{
This function calculates an education sectoral composite score based on various
education-related indicators. It computes scores for disrupted education and
attendance/barriers, combines them into a total score, and determines if a
household is in need or in severe need of educational assistance.

Prerequisite functions:
\itemize{
\item add_loop_age_dummy.R
\item add_loop_edu_access_d.R
\item add_loop_edu_barrier_protection_d.R
\item add_loop_edu_disrupted_d.R
\item add_loop_edu_ind_age_corrected.R
}
}
