% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_edu.R
\name{add_comp_edu}
\alias{add_comp_edu}
\title{Education sectoral composite - add score and dummy for in need}
\usage{
add_comp_edu(
  df,
  ind_schooling_age_d_n = "edu_ind_schooling_age_d_n",
  ind_no_access_d_n = "edu_ind_no_access_d_n",
  ind_barrier_protection_d_n = "ind_barrier_protection_d_n",
  occupation_d_n = "edu_disrupted_occupation_d_n",
  hazards_d_n = "edu_disrupted_hazards_d_n",
  displaced_d_n = "edu_disrupted_displaced_d_n",
  teacher_d_n = "edu_disrupted_teacher_d_n"
)
}
\arguments{
\item{df}{A data frame.}

\item{ind_schooling_age_d_n}{Column name for the number of children of schooling age.}

\item{ind_no_access_d_n}{Column name for the number of children with no access to education.}

\item{ind_barrier_protection_d_n}{Column name for the number of children with barriers to protection.}

\item{occupation_d_n}{Column name for the number of children with disrupted education due to the school being occupied by armed groups}

\item{hazards_d_n}{Column name for the number of children with disrupted education due to hazards.}

\item{displaced_d_n}{Column name for the number of children with disrupted education due to a recent displacement.}

\item{teacher_d_n}{Column name for the number of children with disrupted education due to teachers' absence.}
}
\description{
Education sectoral composite - add score and dummy for in need
}
