% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_edu.R
\name{add_comp_edu}
\alias{add_comp_edu}
\title{Education sectoral composite - add score and dummy for in need}
\usage{
add_comp_edu(
  df,
  schooling_age_n = "edu_schooling_age_n",
  no_access_n = "edu_no_access_n",
  barrier_protection_n = "edu_barrier_protection_n",
  occupation_n = "edu_disrupted_occupation_n",
  hazards_n = "edu_disrupted_hazards_n",
  displaced_n = "edu_disrupted_displaced_n",
  teacher_n = "edu_disrupted_teacher_n"
)
}
\arguments{
\item{df}{A data frame.}

\item{schooling_age_n}{Column name for the number of children of schooling age.}

\item{no_access_n}{Column name for the number of children with no access to education.}

\item{barrier_protection_n}{Column name for the number of children with barriers to protection.}

\item{occupation_n}{Column name for the number of children with disrupted education due to the school being occupied by armed groups}

\item{hazards_n}{Column name for the number of children with disrupted education due to hazards.}

\item{displaced_n}{Column name for the number of children with disrupted education due to a recent displacement.}

\item{teacher_n}{Column name for the number of children with disrupted education due to teachers' absence.}
}
\description{
Education sectoral composite - add score and dummy for in need
}
