% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_foodsec.R
\name{add_comp_foodsec}
\alias{add_comp_foodsec}
\title{Food security sectoral composite - add score and dummy for in need}
\usage{
add_comp_foodsec(
  df,
  fc_phase = "fsl_fc_phase",
  fc_phase_levels = c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC",
    "Phase 5 FC")
)
}
\arguments{
\item{df}{A data frame.}

\item{fc_phase}{Column name for the food security phase.}

\item{fc_phase_levels}{Levels of the food security phase.}
}
\description{
Food security sectoral composite - add score and dummy for in need
}
