% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_prot.R
\name{add_comp_prot}
\alias{add_comp_prot}
\title{Add Protection Composite Score and Need Indicators}
\usage{
add_comp_prot(
  df,
  concern_freq_cope = "prot_concern_freq_cope",
  concern_freq_displaced = "prot_concern_freq_displaced",
  concern_hh_freq_kidnapping = "prot_concern_hh_freq_kidnapping",
  concern_hh_freq_discrimination = "prot_concern_hh_freq_discrimination",
  concern_always = "always",
  concern_several_times = "several_times",
  concern_once_or_twice = "once_or_twice",
  concern_never = "never",
  concern_dnk = "dnk",
  concern_pnta = "pnta"
)
}
\arguments{
\item{df}{A data frame containing the required variables.}

\item{concern_freq_cope}{Column name for concern frequency cope.}

\item{concern_freq_displaced}{Column name for concern frequency displaced.}

\item{concern_hh_freq_kidnapping}{Column name for concern household frequency kidnapping.}

\item{concern_hh_freq_discrimination}{Column name for concern household frequency discrimination.}

\item{concern_always}{Level for always concerned.}

\item{concern_several_times}{Level for concerned several times.}

\item{concern_once_or_twice}{Level for concerned once or twice.}

\item{concern_never}{Level for never concerned.}

\item{concern_dnk}{Level for don't know.}

\item{concern_pnta}{Level for prefer not to answer.}
}
\value{
A data frame with added columns:
\itemize{
\item comp_prot_score_concern_freq_cope: Score for concern frequency cope
\item comp_prot_score_concern_freq_displaced: Score for concern frequency displaced
\item comp_prot_score_concern_hh_freq_kidnapping: Score for concern household frequency kidnapping
\item comp_prot_score_concern_hh_freq_discrimination: Score for concern household frequency discrimination
\item comp_prot_risk_always_d: Indicator for at least one 'Always' response in perceived risks
\item comp_prot_score_concern: Final concern score
\item comp_prot_score: Overall protection composite score
\item comp_prot_in_need: Indicator for protection need
\item comp_prot_in_acute_need: Indicator for acute protection need
}
}
\description{
This function calculates a protection composite score based on concern frequencies. Change for 2025 - child separation categories was
removed.
It also adds indicators for protection needs and acute protection needs.
}
