% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_prot.R
\name{add_comp_prot}
\alias{add_comp_prot}
\title{Protection composite - add score and dummy for in need}
\usage{
add_comp_prot(
  df,
  child_sep_cat = "prot_child_sep_cat",
  child_sep_cat_levels = c("none", "at_least_one_very_severe", "at_least_one_severe",
    "at_least_one_non_severe", "undefined"),
  concern_freq_cope = "prot_concern_freq_cope",
  concern_freq_displaced = "prot_concern_freq_displaced",
  concern_hh_freq_kidnapping = "prot_concern_hh_freq_kidnapping",
  concern_hh_freq_discrimination = "prot_concern_hh_freq_discrimination",
  concern_levels = c("always", "several_times", "once_or_twice", "never", "dnk", "pnta")
)
}
\arguments{
\item{df}{A data frame.}

\item{child_sep_cat}{Column name for child separation category.}

\item{child_sep_cat_levels}{Levels for child separation category.}

\item{concern_freq_cope}{Column name for concern frequency cope.}

\item{concern_freq_displaced}{Column name for concern frequency displaced.}

\item{concern_hh_freq_kidnapping}{Column name for concern household frequency kidnapping.}

\item{concern_hh_freq_discrimination}{Column name for concern household frequency discrimination.}

\item{concern_levels}{Levels for concern variables.}
}
\description{
Protection composite - add score and dummy for in need
}
