% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_snfi.R
\name{add_comp_snfi}
\alias{add_comp_snfi}
\title{Add SNFI Sectoral Composite Score and Need Indicators}
\usage{
add_comp_snfi(
  df,
  shelter_type_cat = "snfi_shelter_type_cat",
  shelter_type_cat_none = "none",
  shelter_type_cat_inadequate = "inadequate",
  shelter_type_cat_adequate = "adequate",
  shelter_type_cat_undefined = "undefined",
  shelter_issue_cat = "snfi_shelter_issue_cat",
  shelter_issue_cat_1_to_3 = "1_to_3",
  shelter_issue_cat_4_to_7 = "4_to_7",
  shelter_issue_cat_8_to_11 = "8_to_11",
  shelter_issue_cat_none = "none",
  shelter_issue_cat_undefined = "undefined",
  shelter_issue_cat_other = "other",
  tenure_security_cat = "hlp_tenure_security",
  tenure_security_cat_high_risk = "high_risk",
  tenure_security_cat_medium_risk = "medium_risk",
  tenure_security_cat_low_risk = "low_risk",
  tenure_security_cat_undefined = "undefined",
  fds_cannot_cat = "snfi_fds_cannot_cat",
  fds_cannot_cat_4 = "4_tasks",
  fds_cannot_cat_2_to_3 = "2_to_3_tasks",
  fds_cannot_cat_1 = "1_task",
  fds_cannot_cat_none = "none",
  fds_cannot_cat_undefined = "undefined",
  shelter_damage_cat = "snfi_shelter_damage_cat",
  shelter_damage_cat_none = "none",
  shelter_damage_cat_damaged = "damaged",
  shelter_damage_cat_part = "part",
  shelter_damage_cat_total = "total",
  shelter_damage_cat_undefined = "undefined"
)
}
\arguments{
\item{df}{A data frame containing the required SNFI indicators.}

\item{shelter_type_cat}{Column name for shelter type.}

\item{shelter_type_cat_none}{Level for no shelter.}

\item{shelter_type_cat_inadequate}{Level for inadequate shelter.}

\item{shelter_type_cat_adequate}{Level for adequate shelter.}

\item{shelter_type_cat_undefined}{Level for undefined shelter.}

\item{shelter_issue_cat}{Column name for shelter issue.}

\item{shelter_issue_cat_1_to_3}{Level for 1 to 3 shelter issues.}

\item{shelter_issue_cat_4_to_7}{Level for 4 to 7 shelter issues.}

\item{shelter_issue_cat_8_to_11}{Level for 8 to 11 shelter issues.}

\item{shelter_issue_cat_none}{Level for no shelter issues.}

\item{shelter_issue_cat_undefined}{Level for undefined shelter issues.}

\item{shelter_issue_cat_other}{Level for other shelter issues.}

\item{tenure_security_cat}{Column name for security of tenure.}

\item{tenure_security_cat_high_risk}{Level for high risk with security of tenure.}

\item{tenure_security_cat_medium_risk}{Level for medium risk with security of tenure.}

\item{tenure_security_cat_low_risk}{Level for low risk with security of tenure.}

\item{tenure_security_cat_undefined}{Level for undefined with security of tenure.}

\item{fds_cannot_cat}{Column name for fds cannot.}

\item{fds_cannot_cat_4}{Level for 4 tasks that cannot be done.}

\item{fds_cannot_cat_2_to_3}{Level for 2 to 3 tasks that cannot be done.}

\item{fds_cannot_cat_1}{Level for 1 task that cannot be done.}

\item{fds_cannot_cat_none}{Level for no tasks that cannot be done.}

\item{fds_cannot_cat_undefined}{Level for undefined fds cannot.}

\item{shelter_damage_cat}{Column for shelter damage category.}

\item{shelter_damage_cat_none}{Level name for no shelter damage.}

\item{shelter_damage_cat_damaged}{Level name for minor damages.}

\item{shelter_damage_cat_part}{Level name for roof with risk of collapse.}

\item{shelter_damage_cat_total}{Level name for total collapse or destruction.}

\item{shelter_damage_cat_undefined}{Level name for undefined or unknown status for shelter damage.}
}
\value{
A data frame with added columns:
\itemize{
\item comp_snfi_score_shelter_type_cat Score based on shelter type
\item comp_snfi_score_shelter_issue_cat Score based on shelter issues
\item comp_snfi_score_tenure_security_cat Score based on security of tenure status
\item comp_snfi_score_fds_cannot_cat Score based on FDS
\item comp_snfi_score_shelter_damage_cat Score based on shelter damage
\item comp_snfi_score: Overall SNFI composite score
\item comp_snfi_in_need: Indicator for being in need
\item comp_snfi_in_severe_need: Indicator for being in severe need
}
}
\description{
This function calculates the Shelter, NFI and HLP (SNFI) sectoral composite score
based on shelter type, shelter issues, occupancy status, shelter damage, and functional
disability scale (FDS) indicators. It also determines if a household is in need or in
severe need based on the calculated score.

Prerequisite functions:
\itemize{
\item add_shelter_issue_cat.R
\item add_shelter_type_cat.R
\item add_occupancy_cat.R
\item add_fds_cannot_cat.R
\item add_shelter_damage_cat.R
}
}
