% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comp_snfi.R
\name{add_comp_snfi}
\alias{add_comp_snfi}
\title{SNFI sectoral composite - add score and dummy for in need}
\usage{
add_comp_snfi(
  df,
  shelter_type_cat = "snfi_shelter_type_cat",
  shelter_type_cat_levels = c("none", "inadequate", "adequate", "undefined"),
  shelter_issue_cat = "snfi_shelter_issue_cat",
  shelter_issue_cat_levels = c("7_to_8", "4_to_6", "1_to_3", "none", "undefined"),
  occupancy_cat = "hlp_occupancy_cat",
  occupancy_cat_levels = c("high_risk", "medium_risk", "low_risk", "undefined"),
  fds_cannot_cat = "snfi_fds_cannot_cat",
  fds_cannot_cat_levels = c("4_to_5_tasks", "2_to_3_tasks", "1_task", "none",
    "undefined")
)
}
\arguments{
\item{df}{A data frame.}

\item{shelter_type_cat}{Column name for shelter type.}

\item{shelter_type_cat_levels}{Levels for shelter type in that order: none, inadequate, adequate, undefined.}

\item{shelter_issue_cat}{Column name for shelter issue.}

\item{shelter_issue_cat_levels}{Levels for shelter issue in that order: none, 7_to_8, 4_to_6, 1_to_3, undefined.}

\item{occupancy_cat}{Column name for occupancy.}

\item{occupancy_cat_levels}{Levels for occupancy in that order: high_risk, medium_risk, low_risk, undefined.}

\item{fds_cannot_cat}{Column name for fds cannot.}

\item{fds_cannot_cat_levels}{Levels for fds cannot. in that order: 4_to_5_tasks, 2_to_3_tasks, 1_task, none, undefined.}
}
\description{
SNFI sectoral composite - add score and dummy for in need
}
