% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drinking_water_source_cat.R
\name{add_drinking_water_source_cat}
\alias{add_drinking_water_source_cat}
\alias{add_drinking_water_time_cat}
\alias{add_drinking_water_time_threshold_cat}
\alias{add_drinking_water_source_jmp_cat}
\title{Drinking water source recoding}
\usage{
add_drinking_water_source_cat(
  df,
  drinking_water_source = "wash_drinking_water_source",
  improved = c("piped_dwelling", "piped_compound", "piped_neighbour", "tap", "borehole",
    "protected_well", "protected_spring", "rainwater_collection", "tank_truck",
    "cart_tank", "kiosk", "bottled_water", "sachet_water"),
  unimproved = c("unprotected_well", "unprotected_spring"),
  surface_water = "surface_water",
  undefined = c("dnk", "pnta", "other")
)

add_drinking_water_time_cat(
  df,
  drinking_water_time_yn = "wash_drinking_water_time_yn",
  water_on_premises = "water_on_premises",
  number_minutes = "number_minutes",
  dnk = "dnk",
  undefined = "pnta",
  drinking_water_time_int = "wash_drinking_water_time_int",
  max = 600,
  drinking_water_time_sl = "wash_drinking_water_time_sl",
  sl_under_30_min = "under_30_min",
  sl_30min_1hr = "30min_1hr",
  sl_more_than_1hr = "more_than_1hr",
  sl_undefined = c("dnk", "pnta"),
  drinking_water_source = "wash_drinking_water_source",
  skipped_drinking_water_source_premises = "piped_dwelling",
  skipped_drinking_water_source_undefined = c("dnk", "pnta")
)

add_drinking_water_time_threshold_cat(
  df,
  drinking_water_time_cat = "wash_drinking_water_time_cat",
  premises = "premises",
  under_30min = c("under_30_min", "30min_1hr"),
  above_30min = "more_than_1hr",
  undefined = "undefined"
)

add_drinking_water_source_jmp_cat(
  df,
  drinking_water_source_cat = "wash_drinking_water_source_cat",
  drinking_water_source_cat_levels = c("improved", "unimproved", "surface_water",
    "undefined"),
  drinking_water_time_30min_cat = "wash_drinking_water_time_30min_cat",
  drinking_water_time_30min_cat_levels = c("premises", "under_30min", "above_30min",
    "undefined")
)
}
\arguments{
\item{df}{A data frame.}

\item{drinking_water_source}{Component column: Water source types.}

\item{improved}{Character vector of responses codes, such as "Protected well" or "Public tap", e.g., c("protected_well", "public_tap").}

\item{unimproved}{Character vector of responses codes, such as "Unprotected well" or "Unprotected spring", e.g., c("unprotected_well", "unprotected_spring").}

\item{surface_water}{Character vector of responses codes, such as "Lake" or "River, e.g., c("lake", "river").}

\item{undefined}{Character vector of responses codes for undefined information, e.g. "Prefer not to answer".}

\item{drinking_water_time_yn}{Component column: Time to fetch water, scoping question.}

\item{water_on_premises}{Character vector of responses codes for water on premises.}

\item{number_minutes}{Character vector of responses codes for number of minutes.}

\item{dnk}{Character vector of responses codes for "Don't know".}

\item{drinking_water_time_int}{Component column: Time to fetch water, integer.}

\item{max}{Integer, the maximum value for the time to fetch water.}

\item{drinking_water_time_sl}{Component column: Time to fetch water, simple choice.}

\item{sl_under_30_min}{Response code for under 30 minutes.}

\item{sl_30min_1hr}{Response code for 30 minutes to 1 hour.}

\item{sl_more_than_1hr}{Response code for more than 1 hour.}

\item{sl_undefined}{Character vector of responses codes for undefined information, e.g. "Don't know" or "Prefer not to answer".}

\item{skipped_drinking_water_source_premises}{Character vector of responses codes for skipped water source on premises, e.g. "Piped into dwelling".}

\item{skipped_drinking_water_source_undefined}{Character vector of responses codes for skipped water source undefined, e.g. "Don't know" or "Prefer not to answer".}

\item{drinking_water_time_cat}{Component column: Time to fetch water, recoded categories.}

\item{premises}{Character vector of responses codes for water on premises.}

\item{under_30min}{Character vector of responses codes for under 30 min.}

\item{above_30min}{Character vector of responses codes for above 30 min.}

\item{drinking_water_source_cat}{Component column: Water source categories.}

\item{drinking_water_source_cat_levels}{Character vector of water source categories in that order: Improved, Unimproved, Surface water, Undefined.}

\item{drinking_water_time_30min_cat}{Component column: Time to fetch water, recoded categories.}

\item{drinking_water_time_30min_cat_levels}{Character vector of time to fetch water categories in that order: Premises, Under 30 min, Above 30 min, Undefined.}
}
\description{
\code{\link[=add_drinking_water_source_cat]{add_drinking_water_source_cat()}} recodes the types of water sources, \code{\link[=add_drinking_water_time_cat]{add_drinking_water_time_cat()}} adds the categorical time to fetch water, and \code{\link[=add_drinking_water_time_threshold_cat]{add_drinking_water_time_threshold_cat()}} according to a chosen threshold. Finally, \link{add_drinking_water_source_jmp_cat} recodes the water source and time to fetch water into a joint JMP category.
}
