% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expenditure_type_freq_rank.R
\name{add_expenditure_type_freq_rank}
\alias{add_expenditure_type_freq_rank}
\title{Rank top 3 frequent expenditure types}
\usage{
add_expenditure_type_freq_rank(
  df,
  expenditure_freq_types = c("cm_expenditure_frequent_food",
    "cm_expenditure_frequent_rent", "cm_expenditure_frequent_water",
    "cm_expenditure_frequent_nfi", "cm_expenditure_frequent_utilities",
    "cm_expenditure_frequent_fuel", "cm_expenditure_frequent_transportation",
    "cm_expenditure_frequent_communication", "cm_expenditure_frequent_other"),
  id_col = "uuid"
)
}
\arguments{
\item{df}{A data frame.}

\item{expenditure_freq_types}{A character vector. The names of the columns that contain the amount of frequent expenditures types.}

\item{id_col}{The name of the column that contains the unique identifier.}
}
\description{
Rank top 3 frequent expenditure types
}
