% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expenditure_type_prop_freq.R
\name{add_expenditure_type_prop_freq}
\alias{add_expenditure_type_prop_freq}
\title{Add frequent expenditure type amount as proportions of total frequent expenditure}
\usage{
add_expenditure_type_prop_freq(
  df,
  cm_expenditure_frequent_food = "cm_expenditure_frequent_food",
  cm_expenditure_frequent_rent = "cm_expenditure_frequent_rent",
  cm_expenditure_frequent_water = "cm_expenditure_frequent_water",
  cm_expenditure_frequent_nfi = "cm_expenditure_frequent_nfi",
  cm_expenditure_frequent_utilitiues = "cm_expenditure_frequent_utilitiues",
  cm_expenditure_frequent_fuel = "cm_expenditure_frequent_fuel",
  cm_expenditure_frequent_transportation = "cm_expenditure_frequent_transportation",
  cm_expenditure_frequent_communication = "cm_expenditure_frequent_communication",
  cm_expenditure_frequent_other = "cm_expenditure_frequent_other"
)
}
\arguments{
\item{df}{A data frame}

\item{cm_expenditure_frequent_food}{Column name for food items expenditure amount.}

\item{cm_expenditure_frequent_rent}{Column name for rent or shelter and/or land expenditure amount.}

\item{cm_expenditure_frequent_water}{Column name for water expenditure amount from all sources combined.}

\item{cm_expenditure_frequent_nfi}{Column name for non-food household items expenditure amount for regular purchase, such as hygiene items, lightbulbs, etc.}

\item{cm_expenditure_frequent_utilitiues}{Column name for utilities expenditure amount, such as electricity or gas connections, etc.}

\item{cm_expenditure_frequent_fuel}{Column name for fuel expenditure amount for cooking, for vehicles, etc.}

\item{cm_expenditure_frequent_transportation}{Column name for transportation expenditure amount, not including vehicle fuel.}

\item{cm_expenditure_frequent_communication}{Column name for communications expenditure amount, such as phone airtime, Internet costs, etc.}

\item{cm_expenditure_frequent_other}{Column name for all other frequent expenditures.}
}
\description{
Add frequent expenditure type amount as proportions of total frequent expenditure
}
