% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expenditure_type_prop_infreq.R
\name{add_expenditure_type_prop_infreq}
\alias{add_expenditure_type_prop_infreq}
\title{Add infrequent expenditure type amount as proportions of total infrequent expenditure}
\usage{
add_expenditure_type_prop_infreq(
  df,
  cm_expenditure_infrequent_shelter = "cm_expenditure_infrequent_shelter",
  cm_expenditure_infrequent_nfi = "cm_expenditure_infrequent_nfi",
  cm_expenditure_infrequent_health = "cm_expenditure_infrequent_health",
  cm_expenditure_infrequent_education = "cm_expenditure_infrequent_education",
  cm_expenditure_infrequent_debt = "cm_expenditure_infrequent_debt",
  cm_expenditure_infrequent_other = "cm_expenditure_infrequent_other"
)
}
\arguments{
\item{df}{A data frame}

\item{cm_expenditure_infrequent_shelter}{Column name for shelter maintenance or repair expenditure amount.}

\item{cm_expenditure_infrequent_nfi}{Column containing expenditure amount for non-food household items for infrequent purchase, such as blankets, cooking pots, clothing, etc.}

\item{cm_expenditure_infrequent_health}{Column name for health-related expenditures, such as healthcare, medicine, etc.}

\item{cm_expenditure_infrequent_education}{Column name for education-related expenditures, such school fees, supplies, uniforms, etc.}

\item{cm_expenditure_infrequent_debt}{Column containing expenditure amount for debt repayment.}

\item{cm_expenditure_infrequent_other}{Column name for all other infrequent expenditures.}
}
\description{
Add infrequent expenditure type amount as proportions of total infrequent expenditure
}
