% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expenditure_type_zero_infreq.R
\name{add_expenditure_type_zero_infreq}
\alias{add_expenditure_type_zero_infreq}
\title{Add zero when the infrequent expenditure type was skipped
Note: Infrequent expenditure = 6-month recall period}
\usage{
add_expenditure_type_zero_infreq(
  df,
  expenditure_infreq = "cm_expenditure_infrequent",
  undefined = c("dnk", "pnta", "none"),
  expenditure_infreq_types = c("cm_expenditure_infrequent_shelter",
    "cm_expenditure_infrequent_nfi", "cm_expenditure_infrequent_health",
    "cm_expenditure_infrequent_education", "cm_expenditure_infrequent_debt",
    "cm_expenditure_infrequent_other")
)
}
\arguments{
\item{df}{A data frame.}

\item{expenditure_infreq}{A character string. The name of the column that contains the infrequent expenditures.}

\item{undefined}{A character vector. The values that indicate that the infrequent expenditures type was skipped.}

\item{expenditure_infreq_types}{A character vector. The names of the columns that contain the amount of infrequent expenditures types.}
}
\description{
Add zero when the infrequent expenditure type was skipped
Note: Infrequent expenditure = 6-month recall period
}
