% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_handwashing_facility_cat.R
\name{add_handwashing_facility_cat}
\alias{add_handwashing_facility_cat}
\title{Add frequent expenditure type amount as proportions of total frequent expenditure}
\usage{
add_handwashing_facility_cat(
  df,
  survey_modality = "survey_modality",
  survey_modality_in_person = c("in_person"),
  survey_modality_remote = c("remote"),
  facility = "wash_handwashing_facility",
  reported_facility = "wash_handwashing_facility_reported",
  observed_soap = "wash_handwashing_facility_observed_soap",
  observed_soap_options = c("soap_available", "soap_not_available",
    "alternative_available"),
  observed_water = "wash_handwashing_facility_observed_water",
  observed_water_options = c("water_available", "water_not_available"),
  facility_options1 = c("available", "none"),
  facility_options2 = c("no_permission"),
  facility_reported_options = c("fixed_dwelling", "fixed_yard", "mobile", "none",
    "other", "dnk", "pnta")
)
}
\arguments{
\item{df}{Data frame containing the survey data.}

\item{survey_modality}{Column name for the survey modality - e.g., "survey_modality".}

\item{survey_modality_in_person}{Options for the in-person survey modality.}

\item{survey_modality_remote}{Options for the remote survey modality.}

\item{facility}{Column name for the observed handwashing facility - e.g., "wash_handwashing_facility".}

\item{reported_facility}{Column name for the reported handwashing facility - e.g., "wash_handwashing_facility_reported".}

\item{observed_soap}{Column name for observed soap availability - e.g., "wash_handwashing_facility_observed_soap".}

\item{observed_soap_options}{Options for the soap availability variable.}

\item{observed_water}{Column name for observed water availability - e.g., "wash_handwashing_facility_observed_water".}

\item{observed_water_options}{Options for the water availability variable.}

\item{facility_options1}{Options for the observed facility variable for in-person surveys.}

\item{facility_options2}{Options for the observed facility variable for remote surveys or "No permission to see".}

\item{facility_reported_options}{Options for the reported facility variable for remote surveys.}
}
\description{
Add frequent expenditure type amount as proportions of total frequent expenditure
}
