% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_handwashing_facility_cat.R
\name{add_handwashing_facility_cat}
\alias{add_handwashing_facility_cat}
\title{Add Handwashing Facility Category}
\usage{
add_handwashing_facility_cat(
  df,
  survey_modality = "survey_modality",
  survey_modality_in_person = c("in_person"),
  survey_modality_remote = c("remote"),
  facility = "wash_handwashing_facility",
  facility_yes = c("available_fixed_in_dwelling", "available_fixed_in_plot",
    "available_fixed_or_mobile"),
  facility_no = "none",
  facility_no_permission = "no_permission",
  facility_undefined = c("other", "pnta"),
  facility_observed_water = "wash_handwashing_facility_observed_water",
  facility_observed_water_yes = "water_available",
  facility_observed_water_no = "water_not_available",
  facility_observed_soap = "wash_soap_observed",
  facility_observed_soap_yes = "yes_soap_shown",
  facility_observed_soap_no = "no",
  facility_observed_undefined = c("other", "pnta"),
  facility_reported = "wash_handwashing_facility_reported",
  facility_reported_yes = c("fixed_dwelling", "fixed_yard", "mobile"),
  facility_reported_no = c("none"),
  facility_reported_undefined = c("other", "dnk", "pnta"),
  facility_reported_water = "wash_handwashing_facility_water_reported",
  facility_reported_water_yes = "yes",
  facility_reported_water_no = "no",
  facility_reported_water_undefined = c("dnk", "pnta"),
  facility_reported_soap = "wash_soap_reported",
  facility_reported_soap_yes = "yes",
  facility_reported_soap_no = "no",
  facility_reported_soap_undefined = c("dnk", "pnta")
)
}
\arguments{
\item{df}{A data frame containing the survey data.}

\item{survey_modality}{Column name for the survey modality (e.g., "survey_modality").}

\item{survey_modality_in_person}{Options for the in-person survey modality.}

\item{survey_modality_remote}{Options for the remote survey modality.}

\item{facility}{Column name for the observed handwashing facility (e.g., "wash_handwashing_facility").}

\item{facility_yes}{Response code indicating availability (e.g., "available").}

\item{facility_no}{Response code indicating unavailability (e.g., "none").}

\item{facility_no_permission}{Response code for cases with no permission to observe.}

\item{facility_undefined}{Response code for undefined situations.}

\item{facility_observed_water}{Column name for observed water availability (e.g., "wash_handwashing_facility_observed_water").}

\item{facility_observed_water_yes}{Response code indicating water is available.}

\item{facility_observed_water_no}{Response code indicating water is not available.}

\item{facility_reported_water}{Column name for reported water availability under no permission or remote conditions.}

\item{facility_reported_water_yes}{Response codes indicating water is available no permission conditions.}

\item{facility_reported_water_no}{Response code indicating water is not reported under no permission or remote conditions.}

\item{facility_reported_water_undefined}{Response codes for undefined water situations under no permission or remote conditions.}

\item{facility_reported_soap}{Column name for reported soap availability under no permission conditions or remote conditions.}

\item{facility_reported_soap_yes}{Response codes indicating soap is reported under no permission or remote conditions.}

\item{facility_reported_soap_no}{Response code indicating soap is not reported under no permission or remote conditions.}

\item{facility_reported_soap_undefined}{Response codes for undefined soap situations under no permission or remote conditions.}

\item{facility_observed_water_undefined}{Response code indicating water is undefined.}
}
\value{
A data frame with an additional column 'wash_handwashing_facility_jmp_cat': Categorized handwashing facilities: "Basic," "Limited," or "No Facility."
}
\description{
This function categorizes handwashing facilities into "Basic," "Limited," or "No facility" based on survey data.
It evaluates various parameters related to the availability of handwashing facilities and the presence of soap and water.
}
