% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_age_dummy.R
\name{add_loop_age_dummy}
\alias{add_loop_age_dummy}
\alias{add_loop_age_dummy_to_main}
\alias{add_loop_age_gender_dummy}
\alias{add_loop_age_gender_dummy_to_main}
\title{Add Age and Gender Dummy Variables for Demographics Loop}
\usage{
add_loop_age_dummy(
  loop,
  ind_age = "ind_age",
  lb = 5,
  ub = 18,
  new_colname = NULL
)

add_loop_age_dummy_to_main(
  main,
  loop,
  ind_age_dummy = "ind_age_5_18",
  id_col_main = "uuid",
  id_col_loop = "uuid",
  new_colname = NULL
)

add_loop_age_gender_dummy(
  loop,
  ind_age = "ind_age",
  lb = 5,
  ub = 18,
  ind_gender = "ind_gender",
  gender = "female",
  new_colname = NULL
)

add_loop_age_gender_dummy_to_main(
  main,
  loop,
  ind_age_gender_dummy = "ind_age_female_5_18",
  id_col_main = "uuid",
  id_col_loop = "uuid",
  new_colname = NULL
)
}
\arguments{
\item{loop}{A data frame of individual-level data.}

\item{ind_age}{Column name for individual age.}

\item{lb}{Lower bound for the age class.}

\item{ub}{Upper bound for the age class.}

\item{new_colname}{New column name for the dummy variable. If NULL, then the default is "ind_age_lb_ub" or "ind_dummy_n" or "ind_age_gender_lb_ub" or "ind_age_gender_dummy_n" depending on the function used.}

\item{main}{A data frame of household-level data.}

\item{ind_age_dummy}{Column name for the dummy variable of the age class.}

\item{id_col_main}{Column name for the unique identifier in the main dataset.}

\item{id_col_loop}{Column name for the unique identifier in the loop dataset.}

\item{ind_gender}{Column name for individual gender.}

\item{gender}{Response options of interest, e.g. "Female".}

\item{ind_age_gender_dummy}{Column name for the dummy variable of the age-gender class.}
}
\value{
Depending on the function used:
\itemize{
\item add_loop_age_dummy: Returns the loop data frame with an additional column for the age dummy variable.
\item add_loop_age_dummy_to_main: Returns the main data frame with an additional column for the sum of age dummy variables.
\item add_loop_age_gender_dummy: Returns the loop data frame with an additional column for the age-gender dummy variable.
\item add_loop_age_gender_dummy_to_main: Returns the main data frame with an additional column for the sum of age-gender dummy variables.
}
}
\description{
This set of functions adds dummy variables for age classes and gender in the demographics loop, and optionally sums these dummies to the main dataset.
It includes four main functions: \code{\link[=add_loop_age_dummy]{add_loop_age_dummy()}}, \code{\link[=add_loop_age_dummy_to_main]{add_loop_age_dummy_to_main()}}, \code{\link[=add_loop_age_gender_dummy]{add_loop_age_gender_dummy()}}, and \code{\link[=add_loop_age_gender_dummy_to_main]{add_loop_age_gender_dummy_to_main()}}.
}
