% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_edu_access_d.R
\name{add_loop_edu_access_d}
\alias{add_loop_edu_access_d}
\alias{add_loop_edu_access_d_to_main}
\title{Add Education Access Dummy Variables}
\usage{
add_loop_edu_access_d(
  loop,
  ind_access = "edu_access",
  yes = "yes",
  no = "no",
  pnta = "pnta",
  dnk = "dnk",
  ind_schooling_age_d = "edu_ind_age_schooling"
)

add_loop_edu_access_d_to_main(
  main,
  loop,
  ind_access_d = "edu_ind_access_d",
  ind_no_access_d = "edu_ind_no_access_d",
  id_col_main = "uuid",
  id_col_loop = "uuid"
)
}
\arguments{
\item{loop}{A data frame of individual-level data for the loop.}

\item{ind_access}{Column name for education access.}

\item{yes}{Value indicating access to education (e.g., "yes").}

\item{no}{Value indicating no access to education (e.g., "no").}

\item{pnta}{Value indicating not applicable (e.g., "pnta").}

\item{dnk}{Value indicating don't know (e.g., "dnk").}

\item{ind_schooling_age_d}{Column name for the dummy variable indicating schooling age.}

\item{main}{A data frame of household-level data.}

\item{ind_access_d}{Column name for education access (binary).}

\item{ind_no_access_d}{Column name for education no access (binary).}

\item{id_col_main}{Column name for the unique identifier in the main dataset.}

\item{id_col_loop}{Column name for the unique identifier in the loop dataset.}
}
\value{
A data frame with additional columns:
\itemize{
\item edu_ind_access_d: Dummy variable indicating access to education (1 if accessed, 0 otherwise).
\item edu_ind_no_access_d: Dummy variable indicating no access to education (1 if not accessed, 0 otherwise).
}
}
\description{
This function adds dummy variables to an individual-level dataframe for all school-aged children.
It creates a variable (\code{edu_ind_access_d}) that indicates whether the child accessed school (1 if accessed, 0 otherwise) and another variable (\code{edu_ind_no_access_d}) that indicates no access to school.

Prerequisite function:
\itemize{
\item add_loop_edu_ind_age_corrected.R
}
}
