% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_edu_access_d.R
\name{add_loop_edu_access_d}
\alias{add_loop_edu_access_d}
\alias{add_loop_edu_access_to_main}
\title{Add education access dummy}
\usage{
add_loop_edu_access_d(
  loop,
  ind_access = "edu_access",
  yes = "yes",
  no = "no",
  pnta = "pnta",
  dnk = "dnk",
  ind_schooling_age_d = "edu_ind_schooling_age_d"
)

add_loop_edu_access_to_main(
  main,
  loop,
  ind_access_d = "edu_ind_access_d",
  ind_no_access_d = "edu_ind_no_access_d",
  id_col_main = "uuid",
  id_col_loop = "uuid"
)
}
\arguments{
\item{loop}{A data frame of individual-level data for the loop.}

\item{ind_access}{Column name for education access.}

\item{yes}{Value for yes..}

\item{no}{Value for no.}

\item{pnta}{Value for not applicable.}

\item{dnk}{Value for don't know.}

\item{ind_schooling_age_d}{Column name for the dummy variable of schooling age.}
}
\description{
This function adds a dummy variable (edu_ind_access_d) to the individual-level dataframe for all school-aged children, taking 1 if the child accessed school and O otherwise. It also adds a no acces variable (edu_ind_no_access_d).
}
