% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_edu_barrier_protection_d.R
\name{add_loop_edu_barrier_protection_d}
\alias{add_loop_edu_barrier_protection_d}
\alias{add_loop_edu_barrier_protection_d_to_main}
\title{Add Child Protection Barriers to Education Variable}
\usage{
add_loop_edu_barrier_protection_d(
  loop,
  barriers = "edu_barrier",
  protection_issues = c("protection_at_school", "protection_travel_school",
    "child_work_home", "child_work_outside", "child_armed_group", "child_marriage",
    "child_pregnancy", "ban", "enroll_lack_documentation", "discrimination"),
  ind_schooling_age_d = "edu_ind_age_schooling"
)

add_loop_edu_barrier_protection_d_to_main(
  main,
  loop,
  ind_barrier_protection_d = "edu_ind_barrier_protection_d",
  id_col_main = "uuid",
  id_col_loop = "uuid"
)
}
\arguments{
\item{loop}{A data frame of individual-level data for the loop.}

\item{barriers}{Column name for the child protection barrier category.}

\item{protection_issues}{Vector of protection issues RESPONSE CODES.}

\item{ind_schooling_age_d}{Column name for the dummy variable of schooling age.}

\item{main}{A data frame of household-level data.}

\item{ind_barrier_protection_d}{Column name for the dummy variable of the child protection barrier category.}

\item{id_col_main}{Column name for the unique identifier in the main dataset.}

\item{id_col_loop}{Column name for the unique identifier in the loop dataset.}
}
\value{
A data frame with an additional column:
\itemize{
\item edu_ind_barrier_protection_d: Dummy variable indicating if a school-aged child faces a protection barrier (1) or not (0).
}

A data frame with an additional column:
\item{edu_barrier_protection_n}{Count of school-aged children facing protection barriers in each household.}
}
\description{
This function adds a dummy variable for child protection barriers to education in individual-level data and aggregates it to household-level data.

Prerequisite function:
\itemize{
\item \code{\link[=add_loop_edu_ind_age_corrected]{add_loop_edu_ind_age_corrected()}}
}
}
