% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_edu_disrupted_d.R
\name{add_loop_edu_disrupted_d}
\alias{add_loop_edu_disrupted_d}
\alias{add_loop_edu_disrupted_d_to_main}
\title{Add education disruption categories}
\usage{
add_loop_edu_disrupted_d(
  df,
  occupation = "edu_disrupted_occupation",
  hazards = "edu_disrupted_hazards",
  displaced = "edu_disrupted_displaced",
  teacher = "edu_disrupted_teacher",
  levels = c("yes", "no", "dnk", "pnta"),
  ind_schooling_age_d = "edu_ind_schooling_age_d"
)

add_loop_edu_disrupted_d_to_main(
  main,
  loop,
  occupation_d = "edu_disrupted_occupation_d",
  hazards_d = "edu_disrupted_hazards_d",
  displaced_d = "edu_disrupted_displaced_d",
  teacher_d = "edu_disrupted_teacher_d",
  id_col_main = "uuid",
  id_col_loop = "uuid"
)
}
\arguments{
\item{df}{A data frame.}

\item{occupation}{Column name for occupation disruption.}

\item{hazards}{Column name for hazards disruption.}

\item{displaced}{Column name for displaced disruption.}

\item{teacher}{Column name for teacher disruption.}

\item{levels}{Vector of levels for the disruption variables.}

\item{ind_schooling_age_d}{Column name for the dummy variable of the schooling age class.}

\item{main}{A data frame of household-level data.}

\item{loop}{A data frame of individual-level data.}

\item{occupation_d}{Column name for the dummy variable of the occupation dimension.}

\item{hazards_d}{Column name for the dummy variable of the hazards dimension.}

\item{displaced_d}{Column name for the dummy variable of the displaced dimension.}

\item{teacher_d}{Column name for the dummy variable of the teacher dimension.}

\item{id_col_main}{Column name for the unique identifier in the main dataset.}

\item{id_col_loop}{Column name for the unique identifier in the loop dataset.}
}
\description{
\code{\link[=add_loop_edu_disrupted_d]{add_loop_edu_disrupted_d()}} adds the education disruption and binaries and \code{\link[=add_loop_edu_disrupted_d_to_main]{add_loop_edu_disrupted_d_to_main()}} adds the education disruption summarised binaries and the category to the main dataset.
}
