% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_edu_ind_age_corrected.R
\name{add_loop_edu_ind_age_corrected}
\alias{add_loop_edu_ind_age_corrected}
\title{Add a correct schooling age to the loop}
\usage{
add_loop_edu_ind_age_corrected(
  loop,
  main,
  id_col_loop = "uuid",
  id_col_main = "uuid",
  survey_start_date = "start",
  school_year_start_month = 9,
  ind_age = "ind_age",
  month = NULL
)
}
\arguments{
\item{loop}{A data frame of individual-level data.}

\item{main}{A data frame of individual-level data.}

\item{id_col_loop}{Survey unique identifier column name in loop.}

\item{id_col_main}{Survey unique identifier column name in main.}

\item{survey_start_date}{Survey start date column name in main.}

\item{school_year_start_month}{The month when the school year has started.}

\item{ind_age}{The individual age column.}

\item{month}{If not NULL, an integer between 1 and 12 which will be used as the month of data collection for all households/}
}
\value{
2 new columns: "edu_ind_age_corrected" with the corrected individual age, and a dummy variable edu_ind_schooling_age_d
}
\description{
Add a correct schooling age to the loop
}
