% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loop_wgq_ss.R
\name{add_loop_wgq_ss}
\alias{add_loop_wgq_ss}
\alias{add_loop_wgq_ss_to_main}
\title{Prepare dummy variables for each WG-SS component (individual data)}
\usage{
add_loop_wgq_ss(
  loop,
  ind_age = "ind_age",
  vision = " wgq_vision",
  hearing = "wgq_hearing",
  mobility = "wgq_mobility",
  cognition = "wgq_cognition",
  self_care = "wgq_self_care",
  communication = "wgq_communication",
  levels = c("no_difficulty", "some_difficulty", "lot_of_difficulty", "cannot_do", "dnk",
    "pnta")
)

add_loop_wgq_ss_to_main(
  main,
  loop,
  wgq_dis = "wgq_dis",
  ind_age_above_5 = "ind_age_above_5",
  id_col_main = "uuid",
  id_col_loop = "uuid"
)
}
\arguments{
\item{loop}{A data frame of individual-level data.}

\item{ind_age}{The individual age column.}

\item{vision}{Vision component column.}

\item{hearing}{Hearing component column.}

\item{mobility}{Mobility component column.}

\item{cognition}{Cognition component column.}

\item{self_care}{Self-care component column.}

\item{communication}{Communication component column.}

\item{levels}{Character vector of responses codes, including first in the following order: "No difficulty", "Some difficulty", "A lot of difficulty", "Cannot do at all", e.g. c("no_difficulty", "some_difficulty", "lot_of_difficulty", "cannot_do").}

\item{main}{A data frame of household-level data.}

\item{wgq_dis}{Column name for the disability dummy in the individual-level dataset.}

\item{ind_age_above_5}{Column name for the age above 5 dummy in the individual-level dataset.}

\item{id_col_main}{Column name for the unique identifier in the main dataset.}

\item{id_col_loop}{Column name for the unique identifier in the loop dataset.}
}
\value{
Fifteen new columns: each component dummy for levels 3 and 4, disability level 3, level 4, and level 3 and 4 together.
}
\description{
Prepare dummy variables for each WG-SS component (individual data)
}
