% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_msni.R
\name{add_msni}
\alias{add_msni}
\title{Add MSNI - add score and dummy for in need}
\usage{
add_msni(
  df,
  comp_foodsec_score = "comp_foodsec_score",
  comp_snfi_score = "comp_snfi_score",
  comp_wash_score = "comp_wash_score",
  comp_prot_score = "comp_prot_score",
  comp_health_score = "comp_health_score",
  comp_edu_score = "comp_edu_score"
)
}
\arguments{
\item{df}{A data frame.}

\item{comp_foodsec_score}{Column name for the food security composite score.}

\item{comp_snfi_score}{Column name for the SNFI composite score.}

\item{comp_wash_score}{Column name for the WASH composite score.}

\item{comp_prot_score}{Column name for the protection composite score.}

\item{comp_health_score}{Column name for the health composite score.}

\item{comp_edu_score}{Column name for the education composite score.}
}
\description{
The output is a data frame with three new columns: \code{msni_score}, "msni_in_need", and "msni_in_acute_need".`
}
