% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_occupancy_cat.R
\name{add_occupancy_cat}
\alias{add_occupancy_cat}
\title{Add the category of occupancy arrangement}
\usage{
add_occupancy_cat(
  df,
  occupancy = "hlp_occupancy",
  high_risk = c("no_agreement"),
  medium_risk = c("rented", "hosted_free"),
  low_risk = c("ownership"),
  undefined = c("dnk", "pnta", "other")
)
}
\arguments{
\item{df}{A data frame.}

\item{occupancy}{Component column: Occupancy arrangement.}

\item{high_risk}{Character vector of high risk occupancy arrangements.}

\item{medium_risk}{Character vector of medium risk occupancy arrangements.}

\item{low_risk}{Character vector of low risk occupancy arrangements.}

\item{undefined}{Character vector of undefined responses codes (e.g. "Prefer not to answer").}
}
\description{
Add the category of occupancy arrangement
}
