% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quantile_interval.R
\name{add_quantile_interval}
\alias{add_quantile_interval}
\title{Add quantile intervals}
\usage{
add_quantile_interval(
  df,
  vars,
  cut_offs = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  weight = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{The variable or variables with income integer values.}

\item{cut_offs}{The cut-offs for computaton of quantiles. The default corresponds to the quantiles of the distribution, based on the weighted median.}

\item{weight}{The sampling weights.}
}
\description{
\code{\link[=add_quantile_interval]{add_quantile_interval()}} function calculates weighted income quantiles and classifies households according to these.
}
