% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sanitation_facility_cat.R
\name{add_sanitation_facility_cat}
\alias{add_sanitation_facility_cat}
\alias{add_sharing_sanitation_facility_cat}
\alias{add_sharing_sanitation_facility_n_ind}
\alias{add_sanitation_facility_jmp_cat}
\title{Sanitation facility classification}
\usage{
add_sanitation_facility_cat(
  df,
  sanitation_facility = "wash_sanitation_facility",
  improved = c("flush_piped_sewer", "flush_septic_tank", "flush_pit_latrine",
    "flush_dnk_where", "pit_latrine_slab", "twin_pit_latrine_slab",
    "ventilated_pit_latrine_slab", "container", "compost"),
  unimproved = c("flush_open_drain", "flush_elsewhere", "pit_latrine_wo_slab", "bucket",
    "hanging_toilet", "plastic_bag"),
  none = "none",
  undefined = c("other", "dnk", "pnta")
)

add_sharing_sanitation_facility_cat(
  df,
  sharing_sanitation_facility = "wash_sanitation_facility_sharing_yn",
  yes = "yes",
  no = "no",
  undefined = c("dnk", "pnta"),
  sanitation_facility = "wash_sanitation_facility",
  skipped_sanitation_facility = NULL
)

add_sharing_sanitation_facility_n_ind(
  df,
  sharing_sanitation_facility_cat = "wash_sharing_sanitation_facility_cat",
  levels = c("shared", "not_shared", "not_applicable", "undefined"),
  sanitation_facility_sharing_n = "wash_sanitation_facility_sharing_n",
  hh_size = "hh_size",
  weight = "weight"
)

add_sanitation_facility_jmp_cat(
  df,
  sanitation_facility_cat = "wash_sanitation_facility_cat",
  sanitation_facility_levels = c("improved", "unimproved", "none", "undefined"),
  sharing_sanitation_facility_cat = "wash_sharing_sanitation_facility_cat",
  sharing_sanitation_facility_levels = c("shared", "not_shared", "not_applicable",
    "undefined")
)
}
\arguments{
\item{df}{A data frame.}

\item{sanitation_facility}{Component column: Sanitation facility types.}

\item{improved}{Character vector of responses codes for Improved facilities.}

\item{unimproved}{Character vector of responses codes for Unimproved facilities.}

\item{none}{Character vector of responses codes for No sanitation facility/Open defecation.}

\item{undefined}{Character vector of responses codes, that do not fit any category, e.g., c("dnk", "pnta", "other").}

\item{sharing_sanitation_facility}{Component column: Number of people with whom the facility is shared.}

\item{yes}{Character vector of responses codes for Yes.}

\item{no}{Character vector of responses codes for No.}

\item{skipped_sanitation_facility}{Character vector of responses codes for skipped sanitation facility.}

\item{sharing_sanitation_facility_cat}{Component column: Sharing status of sanitation facility recoded.}

\item{levels}{Character vector of responses codes, including first in the following order: Shared, Not shared, Not applicable, and Undefined.}

\item{sanitation_facility_sharing_n}{Component column: number of households sharing the sanitation facility.}

\item{hh_size}{Column of the household size.}

\item{weight}{Column of the survey weights.}

\item{sanitation_facility_cat}{Component column: Sanitation facility types recoded.}

\item{sanitation_facility_levels}{Levels: Improved, Unimproved, None, Undefined.}

\item{sharing_sanitation_facility_levels}{Levels: Shared, Not shared, Not applicable, Undefined.}
}
\description{
\code{\link[=add_sanitation_facility_cat]{add_sanitation_facility_cat()}} recodes the types of sanitation facilities, \code{\link[=add_sharing_sanitation_facility_cat]{add_sharing_sanitation_facility_cat()}} recodes the sharing status of sanitation facility, and \code{\link[=add_sanitation_facility_jmp_cat]{add_sanitation_facility_jmp_cat()}} combines the previous two functions to recode the sanitation facility into a JMP classification. Finally, \code{\link[=add_sharing_sanitation_facility_num_ind]{add_sharing_sanitation_facility_num_ind()}} recodes the number of individuals sharing the sanitation facility.
}
