% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shelter_damage_cat.R
\name{add_shelter_damage_cat}
\alias{add_shelter_damage_cat}
\title{Add Category of Shelter damage (Optional SNFI dimension)}
\usage{
add_shelter_damage_cat(
  df,
  snfi_shelter_damage = "snfi_shelter_damage",
  snfi_shelter_damage_none = "none",
  snfi_shelter_damage_damaged = c("minor", "damage_windows_doors", "damage_floors",
    "damage_walls"),
  snfi_shelter_damage_part = "major",
  snfi_shelter_damage_total = "total_collapse",
  snfi_shelter_damage_undefined = c("dnk", "pnta", "other")
)
}
\arguments{
\item{df}{A data frame containing shelter damage arrangement.}

\item{snfi_shelter_damage}{The name of the column  containing shelter damage status.}

\item{snfi_shelter_damage_none}{The value indicating no damage.}

\item{snfi_shelter_damage_damaged}{Values indicating minor damage to roof and other damages.}

\item{snfi_shelter_damage_part}{The value indicating major damage to roof with risk of collapse.}

\item{snfi_shelter_damage_total}{The value indicating total collapse or destruction.}

\item{snfi_shelter_damage_undefined}{Values indicating undefined or unknown status.}
}
\value{
A data frame with an additional column:
\itemize{
\item snfi_shelter_damage_cat: Categorized shelter damages: "none", "damaged", "part", "total", or "undefined".
}
}
\description{
This function categorizes No damage, damaged, partial collapse / partial destruction based on provided criteria.
}
