% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shelter_issue_cat.R
\name{add_shelter_issue_cat}
\alias{add_shelter_issue_cat}
\title{Add the number of shelter issues and related category}
\usage{
add_shelter_issue_cat(
  df,
  shelter_issue = "snfi_shelter_issue",
  none = "none",
  issues = c("lack_privacy", "lack_space", "temperature", "ventilation", "leak", "lock",
    "lack_lighting", "difficulty_move"),
  undefined = c("dnk", "pnta", "other"),
  sep = "/"
)
}
\arguments{
\item{df}{A data frame.}

\item{shelter_issue}{Component column: Shelter issues.}

\item{none}{Response code for no issue.}

\item{issues}{Character vector of issues.}

\item{undefined}{Character vector of undefined responses codes (e.g. "Prefer not to answer").}

\item{sep}{Separator for the dummy columns.}
}
\description{
Add the number of shelter issues and related category
}
