% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{are_values_in_set}
\alias{are_values_in_set}
\title{Stop statement values are not in set}
\usage{
are_values_in_set(
  df,
  cols,
  set,
  main_message = "All columns must be in the following set: "
)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}

\item{set}{A vector of values}

\item{main_message}{A main message}
}
\value{
A stop statement
}
\description{
Stop statement values are not in set
}
