% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{are_values_in_set}
\alias{are_values_in_set}
\title{Stop statement values are not in set}
\usage{
are_values_in_set(df, cols, set)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}

\item{set}{A vector of values}
}
\value{
A stop statement
}
\description{
Stop statement values are not in set
}
