% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_if_above_zero.R
\name{count_if_above_zero}
\alias{count_if_above_zero}
\title{Create dummy variables and count across if above zero}
\usage{
count_if_above_zero(df, vars, new_colname)
}
\arguments{
\item{df}{A data frame.}

\item{vars}{Vector of column names.}

\item{new_colname}{The name of the new column.}
}
\description{
Create dummy variables and count across if above zero
}
