% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_source.R
\name{drinking_water_source}
\alias{drinking_water_source}
\alias{time_to_fetch_water}
\alias{drinking_water_source_score}
\title{Water source classification - 5-point scale}
\usage{
drinking_water_source(
  df,
  drinking_water_source = "drinking_water_source",
  improved = c("protected_well", "public_tap"),
  unimproved = c("unprotected_well", "unprotected_spring"),
  surface_water = c("lake", "river"),
  na = c("other")
)

time_to_fetch_water(
  df,
  time_to_fetch_water = "time_to_fetch_water",
  above_threshold = c("30mins_to_59mins", "1hour_above"),
  below_threshold = c("5mins_to_14mins", "15mins_to_29mins"),
  premises = c("premises"),
  na = c("other")
)

drinking_water_source_score(
  df,
  drinking_water_source_cat = "drinking_water_source_cat",
  drinking_water_source_levels = c("improved", "unimproved", "surface_water"),
  time_to_fetch_water_cat = "time_to_fetch_water_cat",
  time_to_fetch_water_levels = c("premises", "below_threshold", "above_threshold")
)
}
\arguments{
\item{df}{A data frame.}

\item{drinking_water_source}{Component column: Water source types.}

\item{improved}{Character vector of responses codes, such as "Protected well" or "Public tap", e.g., c("protected_well", "public_tap").}

\item{unimproved}{Character vector of responses codes, such as "Unprotected well" or "Unprotected spring", e.g., c("unprotected_well", "unprotected_spring").}

\item{surface_water}{Character vector of responses codes, such as "Lake" or "River, e.g., c("lake", "river").}

\item{na}{Character vector of responses codes, that do not fit any category, e.g., c("other").}

\item{above_threshold}{Character vector of responses codes, such as "30 minutes to 59 minutes" or "1 hour and above", e.g., c("30mins_to_59mins", "1hour_above").}

\item{below_threshold}{Character vector of responses codes, such as "5 to 14 minutes" or "15 minutes to 29 minutes", e.g., c("5mins_to_14mins", "15mins_to_29mins").}

\item{premises}{Character vector of responses codes, such as "On premises", e.g., c("premises").}

\item{drinking_water_source_cat}{Drinking water sources column.}

\item{drinking_water_source_levels}{Water sources levels - in that order: improved, unimproved, surface water.}

\item{time_to_fetch_water_cat}{Time to fetch drinking water column.}

\item{time_to_fetch_water_levels}{Time to fetch drinking water levels - in that order: premises, below the threshold, above the threshold.}

\item{time_to_fetch}{Component column: Time to fetch water.}
}
\value{
A dataframe with a new column. For \code{\link[=drinking_water_source]{drinking_water_source()}}, a recoded column of water sources between improved, unimproved and surface water (drinking_water_source_cat); for \code{\link[=time_to_fetch_water]{time_to_fetch_water()}}, a recoded column of times to fetch water according to the chosen thresholds (time_to_fetch_water_cat); for \code{\link[=drinking_water_source_score]{drinking_water_source_score()}}, a 5-point scale from 1 to 5 (drinking_water_source_score).
}
\description{
\code{\link[=drinking_water_source]{drinking_water_source()}} recodes the types of water sources, \code{\link[=time_to_fetch_water]{time_to_fetch_water()}} the time to fetch water according to a chosen threshold, and \code{\link[=drinking_water_source_score]{drinking_water_source_score()}} classify each household/individual on a 5-point scale.
}
\section{Details on the 5-point scale}{


The classification is computed as follows:
\itemize{
\item Level 5: Surface water;
\item Level 4: Unimproved water source;
\item Level 3: Improved water source with more than 30 minutes return time;
\item Level 2: Improved water source within 30 minutes return time;
\item Level 1: Improved water source on premises.
}
}

