% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humind-package.R
\docType{package}
\name{humind-package}
\alias{humind}
\alias{humind-package}
\title{humind: Compose MSNA Indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This packages provides functions to compose usual humanitarian composite indicators related to IMPACT Multi-Sector Needs Assessments (MSNAs) such as food sec indicators (rCSI, LCSI, HHS, FCS), Washington Group Short Set of questions, JMP ladders, as well as specific indicators related to the MSNI framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/impact-initiatives-hppu/humind}
  \item \url{https://impact-initiatives-hppu.github.io/humind}
  \item Report bugs at \url{https://github.com/impact-initiatives-hppu/humind/issues}
}

}
\author{
\strong{Maintainer}: Quentin VILLOTTA \email{quentin.villotta@impact-initiatives.org}

Authors:
\itemize{
  \item Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org} [contributor]
  \item Mario DASILVA \email{mario.fidalgo@impact-initiatives.org} [contributor]
}

Other contributors:
\itemize{
  \item Laith Abu-Nawwas \email{laith.abu-nawwas@impact-initiatives.org} [contributor]
  \item Sunhee Moraes \email{sunhee.moraes@impact-initiatives.org} [contributor]
  \item Raphael Bacot \email{raphael.bacot@impact-initiatives.org} [contributor]
}

}
\keyword{internal}
