% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humind-package.R
\docType{package}
\name{humind-package}
\alias{humind}
\alias{humind-package}
\title{humind: Compose MSNA Indicators}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This packages provides functions to compose usual humanitarian composite indicators related to IMPACT Multi-Sector Needs Assessments (MSNAs) such as food sec indicators (rCSI, LCSI, HHS, FCS), Washington Group Short Set of questions, JMP ladders, as well as specific indicators related to the MSNI framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives-hppu.github.io/humind}
  \item \url{https://github.com/impact-initiatives-hppu/humind}
  \item Report bugs at \url{https://github.com/impact-initiatives-hppu/humind/issues}
}

}
\author{
\strong{Maintainer}: Quentin VILLOTTA \email{quentin.villotta@impact-initiatives.org}

Authors:
\itemize{
  \item Quentin VILLOTTA \email{quentin.villotta@impact-initiatives.org}
  \item Guillaume Noblet
  \item Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}
  \item Mario Fidalgo \email{mario.fidalgo@impact-initiatives.org}
}

Other contributors:
\itemize{
  \item Laith Abu-Nawwas [contributor]
  \item Sunhee Moraes [contributor]
  \item Raphael Bacot [contributor]
  \item IMPACT Initiatives [copyright holder]
}

}
\keyword{internal}
