% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humind-package.R
\docType{package}
\name{humind-package}
\alias{humind}
\alias{humind-package}
\title{humind: Calculate Usual Humanitarian Composite Indicators}
\description{
Give helpers to calculate usual humanitarian composite indicators such as food sec indicators (rCSI, LCSI, HHS, FCS), Washington Group Short Set of questions, HWISE scales, recode water sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/impact-initiatives-hppu/humind}
  \item \url{https://gnoblet.github.io/impact-initiatives-hppu/}
  \item Report bugs at \url{https://github.com/impact-initiatives-hppu/humind/issues}
}

}
\author{
\strong{Maintainer}: Guillaume Noblet \email{gnoblet@zaclys.net}

}
\keyword{internal}
