% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{if_not_in_stop}
\alias{if_not_in_stop}
\title{Stop statement "If not in colnames" with colnames}
\usage{
if_not_in_stop(df, cols, df_name, arg = NULL)
}
\arguments{
\item{df}{A data frame}

\item{cols}{A vector of column names (quoted)}

\item{df_name}{Provide the tibble name as a character string}

\item{arg}{Default to NULL.}
}
\value{
A stop statement
}
\description{
Stop statement "If not in colnames" with colnames
}
