% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_value.R
\name{impute_value}
\alias{impute_value}
\alias{impute_median}
\title{Impute Missing Values}
\usage{
impute_value(df, vars, value)

impute_median(df, vars, group = NULL, weighted = FALSE, weight = NULL)
}
\arguments{
\item{df}{A dataframe containing the variables to be imputed.}

\item{vars}{A character vector of the variables to replace NA with.}

\item{value}{The value to replace NA with.}

\item{group}{A character vector of the grouping variables.}

\item{weighted}{A boolean indicating whether to use the weighted median or not.}

\item{weight}{The weight variable.}
}
\value{
A dataframe with imputed values:
\item{vars}{The specified variables with missing values replaced}

A dataframe with imputed values:
\item{vars}{The specified variables with missing values replaced by their respective (weighted) medians}
}
\description{
This set of functions replaces missing values in specified variables.
\code{\link[=impute_value]{impute_value()}} replaces all missing values with a specific value, while
\code{\link[=impute_median]{impute_median()}} replaces missing values with the (weighted) median of the variable.
}
