% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is_df}
\alias{is_df}
\title{Stop statement "If input is not a data.frame"}
\usage{
is_df(df, arg = NULL)
}
\arguments{
\item{df}{A data frame}

\item{arg}{Default to NULL.}
}
\value{
\code{invisible(TRUE)} if \code{df} is a data frame, otherwise a stop statement.
}
\description{
Stop statement "If input is not a data.frame"
}
