% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_cat.R
\name{num_cat}
\alias{num_cat}
\title{Add Categories for a Numeric Variable}
\usage{
num_cat(
  df,
  num_col,
  breaks,
  labels = NULL,
  int_undefined = c(-999, 999),
  char_undefined = "Unknown",
  new_colname = NULL,
  plus_last = FALSE,
  above_last = FALSE
)
}
\arguments{
\item{df}{A dataframe containing the numeric column to be categorized.}

\item{num_col}{The column name to recategorize.}

\item{breaks}{A vector of cut points for categorization.}

\item{labels}{A vector of labels for the categories. If NULL, labels are automatically generated.}

\item{int_undefined}{A vector of numeric values to be replaced by char_undefined.}

\item{char_undefined}{A character string to replace int_undefined values.}

\item{new_colname}{The name of the new column. If NULL, it's automatically generated.}

\item{plus_last}{Logical, whether to add a "+" to the last category.}

\item{above_last}{Logical, whether to add a category for values above the last break.}
}
\value{
A dataframe with an additional column:
\itemize{
\item new_colname: A new column containing the categorized values of the original numeric column.
}
}
\description{
This function categorizes a numeric column in a dataframe based on specified breaks and labels. It can handle undefined values and provides options for customizing the categorization process.
}
