% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_top3_vars.R
\name{rank_top3_vars}
\alias{rank_top3_vars}
\title{Rank Top 3 Variables}
\usage{
rank_top3_vars(
  df,
  vars,
  new_colname_top1,
  new_colname_top2,
  new_colname_top3,
  id_col = "uuid"
)
}
\arguments{
\item{df}{A data frame containing the variables to be ranked.}

\item{vars}{A character vector of numeric variable names to be ranked.}

\item{new_colname_top1}{The new column name for the top variable.}

\item{new_colname_top2}{The new column name for the 2nd top variable.}

\item{new_colname_top3}{The new column name for the 3rd top variable.}

\item{id_col}{The column name for the unique identifier.}
}
\value{
A data frame with additional columns:
\itemize{
\item new_colname_top1: The name of the top-ranked variable for each row.
\item new_colname_top2: The name of the second-ranked variable for each row.
\item new_colname_top3: The name of the third-ranked variable for each row.
}
}
\description{
This function ranks the top 3 numeric variables for each row in a data frame. In case of equality, it takes precedence in the order of 'vars' as factors.
}
