% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitation_facility.R
\name{sanitation_facility}
\alias{sanitation_facility}
\title{Sanitation facility classification - 5-point scale}
\usage{
sanitation_facility(
  df,
  sanitation_facility = "sanitation_facility",
  sanitation_facility_improved_codes = c("composting toilet", "pour_flush_toilet"),
  sanitation_facility_unimproved_codes = c("bucket", "hanging_latrine"),
  sanitation_facility_open_defecation_codes = c("open_defecation"),
  sanitation_facility_na_codes = c("other"),
  sharing_sanitation_facility = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{sanitation_facility}{Component column: Sanitation facility types.}

\item{sanitation_facility_improved_codes}{Character vector of responses codes, such as "Composting toilet" or "Pour flush toilet", e.g., c("composting toilet", "pour_flush_toilet").}

\item{sanitation_facility_unimproved_codes}{Character vector of responses codes, such as "Bucket" or "Hanging latrine", e.g., c("bucket", "hanging_latrine").}

\item{sanitation_facility_open_defecation_codes}{Character vector of responses codes, such as "Open defecation", e.g., c("open_defecation").}

\item{sanitation_facility_na_codes}{Character vector of responses codes, that do not fit any category, e.g., c("other").}

\item{sharing_sanitation_facility}{Component column: Number of people with whom the facility is shared.}
}
\value{
Three new columns: a recoded column of sanitation facilities between improved, unimproved and open defecatop,my finger  (water_source_recoded), a recoded column of times to fetch water according to the chosen thresholds (time_to_fetch_recoded), a 5-point scale from 1 to 5 (water_source_class).
}
\description{
\code{sanitation_facility()} recodes the types of sanitation facilities, and, if present, the number of persons sharing the facility, and classify each household/individual on a 5-point scale.
}
\section{Details on the 5-point scale}{


The classification is computed as follows:
\itemize{
\item Level 5: Open defecation;
\item Level 4: Unimproved sanitation facility;
\item Level 3: Improved sanitation facility with more than 50 people;
\item Level 2: Improved sanitation facility with 20 to 49 people;
\item Level 1: Improved sanitation facility with less than 19 people.
}

If \code{sharing_sanitation_facility} is null, then levels 4 and 5 do not change, and improved sanitation facilities assign a 1.
}

