% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_vars.R
\name{sum_vars}
\alias{sum_vars}
\title{Sum Columns Row-wise}
\usage{
sum_vars(
  df,
  vars,
  new_colname,
  imputation = "none",
  na_rm = FALSE,
  weight = "weight",
  value = 0,
  group = NULL
)
}
\arguments{
\item{df}{A dataframe containing the columns to be summed.}

\item{vars}{A character vector of the columns to sum.}

\item{new_colname}{A character vector of the new column name.}

\item{imputation}{The imputation method, either none (default), value, median or weighted median.}

\item{na_rm}{A boolean indicating whether to remove missing values when summing across columns.}

\item{weight}{The weight variable to calculate weighted means or medians.}

\item{value}{The value to replace missing values with if imputation is "value".}

\item{group}{A character vector of the grouping variables.}
}
\value{
A dataframe with an additional column:
\itemize{
\item new_colname: the sum of the specified variables, with imputation applied if specified.
}
}
\description{
This function sums up specified columns row-wise in a dataframe, with options for imputation of missing values and weighted calculations.
}
